% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{ap}
\alias{ap}
\title{Airports}
\format{a \code{dataframe} instance, 1 row per station with columns:
\itemize{
\item 'name':    A \code{character}  Name of airport. May or may not contain the City name.
\item 'city':    A \code{character}  Main city served by airport. May be spelled differently from Name.
\item 'country': A \code{character}  Country or territory where airport is located.
\item 'IATA':    A \code{character}  3-letter IATA code
\item 'ICAO':    A \code{numeric}    4-letter ICAO code
\item 'lat':     A \code{numeric}    Latitude of airport
\item 'lon':     A \code{numeric}    Longitude of airport
}}
\source{
\href{https://openflights.org/data.html}{OpenFlights}
}
\usage{
ap
}
\description{
Dataset containing information global Airports subset to the Daymet Domain
}
\examples{
\dontrun{
 airports = HydroData::ap
}

}
\keyword{datasets}
