% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.merge.R
\name{clip.merge}
\alias{clip.merge}
\title{Clip and merge raster data}
\usage{
clip.merge(dataset, bounds)
}
\arguments{
\item{dataset}{a list of rasters}

\item{bounds}{a boundary to clip to}
}
\description{
Clip raster data to a set boundary and merge
}
\author{
Mike Johnson
}
