% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipMerge.R
\name{clipMerge}
\alias{clipMerge}
\title{Clip and merge raster data}
\usage{
clipMerge(dataset, bounds)
}
\arguments{
\item{dataset}{a list of rasters}

\item{bounds}{a boundary to clip to}
}
\description{
Clip raster data to a set boundary and merge
}
\author{
Mike Johnson
}
