% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.palletes.R
\docType{data}
\name{col_elev}
\alias{col_elev}
\title{Color Palletes Usefull for NLCD, CDL and NED plotting}
\format{An object of class \code{character} of length 23.}
\usage{
col_elev
}
\description{
Color pallets for visualizing land use and elevation raster data
}
\author{
Mike Johnson
}
\keyword{datasets}
