% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_AOI.R
\name{define_AOI}
\alias{define_AOI}
\title{Define Area of interest (AOI)}
\usage{
define_AOI(state = NULL, county = NULL, clip_unit = NULL,
  get.basemap = FALSE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{get_basemap}{logical. If TRUE AOI shapefile will be returned as a list also containing a basemap of AOI}
}
\description{
Internal function used to generate AOI shapefiles in all HydroData functions.
}
\examples{
#By state
define_AOI(state = 'CA')

#By state, county combination
define_AOI(state = 'California', county = 'Santa Barbara')

#By external shapefile
define_AOI(clip_unit = rgdal::readOGR('la_metro.shp'))

#By 10 mile2 bounding box using users location as centroid
define_AOI(clip_unit = c(get_ip_loc(), 10, 10))

#By 10 mile2 bounding box using the 'KMART near UCSB' as centroid
define_AOI(clip_unit = c('KMART near UCSB', 10, 10))

#By HUC8 unit covering users location
define_AOI(clip_unit = get_WBD(get_ip_loc(), level = 8))
}
\author{
Mike Johnson
}
