% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_download_url.R
\name{download.url}
\alias{download.url}
\title{Download data from URL}
\usage{
download.url(url, mode = "curl", max.try = 5)
}
\arguments{
\item{url}{the url}

\item{mode}{download mode}

\item{max.try}{number of attempts allowed}
}
\description{
Download data from URL
}
