% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_download_shp.R
\name{download_shp}
\alias{download_shp}
\title{Download, unzip and read shapefile from URL}
\usage{
download_shp(URL, type)
}
\arguments{
\item{URL}{path to data}

\item{type}{description of data being downloaded used for messaging#'}
}
\value{
a \code{Spatial} object retrived from URL and projected to "+init=epsg:4326"
}
\description{
Download, unzip and read shapefile from URL
}
\examples{
\dontrun{
download.shp(URL, type = "flowlines")
}

}
\seealso{
Other HydroData 'utility' function: \code{\link{getBoundingBox}}
}
\author{
Mike Johnson
}
