% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore Spatial data found with HydroData}
\usage{
explore(input = NULL, save = FALSE)
}
\description{
\code{explore} can be used to explore timeseries HydroData objects
Namely:
 findGHCN
 findNHD
 findReservoir
 findRoads
 findSnotel
 findUSGS
 findWaterbodies
 findWS
 findAirports
}
\details{
getCLD
 getKoppen
 getNED
 getNLCD
 getWaterUse
}
