% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flowlines_1.R
\name{find_flowlines}
\alias{find_flowlines}
\title{Find all NHD flowlines within an Area of Interest}
\usage{
find_flowlines(state = NULL, county = NULL, clip_unit = NULL,
  keep.boundary = FALSE, keep.basemap = FALSE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with gage data in a list}

\item{keep.basemap}{logical. If TRUE, the google basemap will be returned with gage data in a list}
}
\description{
Function to locate all NHD flowlines within an Area of Interest
}
\examples{
Find all NHD flowlines in El Paso County, Colorado

el.paso = find_flowlines(state = 'CO', county = 'El Paso', keep.boundary = TRUE, keep.basemap = TRUE)

plot(el.paso$basemap)
plot(el.paso$boundary, add = TRUE, lwd = 5)
plot(el.paso$flowlines, add = TRUE, col = 'blue', lwd = el.paso$flowlines$streamorde)

Get NHD COMIDs

comids = el.paso$flowlines$comid

}
\author{
Mike Johnson
}
