% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNID.R
\name{find_nid}
\alias{find_nid}
\title{Locate all USACE NID Dams within Area of Interest}
\usage{
find_nid(state = NULL, county = NULL, clip_unit = NULL,
  keep.boundary = FALSE, keep.basemap = FALSE, save = FALSE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with gage data in a list}

\item{keep.basemap}{logical. If TRUE, the google basemap will be returned with gage data in a list}

\item{save}{logical. If TRUE, all data is written to a HydroData folder in the working directory}
}
\description{
Function to locate all US Army Corps dams within an Area of Interest from the National Inventory of Dams
}
\examples{
Find all dams in Texas

tx.dams = find_nid(state = "TX", keep.boundary = TRUE, keep.basemap = TRUE, save= TRUE)
plot(tx.dams$basmap)
plot(tx.dams$boundary, add = TRUE)
plot(tx.dams$dams, add = TRUE)

}
\author{
Mike Johnson
}
