% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reservoir.R
\name{find_res}
\alias{find_res}
\title{Locate all USACE NID Dams within Area of Interest}
\usage{
find_res(state = NULL, county = NULL, clip_unit = NULL,
  keep.boundary = FALSE, keep.basemap = FALSE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with gage data in a list}

\item{keep.basemap}{logical. If TRUE, the google basemap will be returned with gage data in a list}
}
\description{
Function to locate all US Army Corps dams within an Area of Interest from the National Inventory of Dams
}
\examples{
Find all reservoirs in Texas

tx.res = find_reservoir(state = "TX", keep.boundary = TRUE, keep.basemap = TRUE)
plot(tx.res$basmap)
plot(tx.res$boundary, add = TRUE)
plot(tx.res$dams, add = TRUE)

}
\author{
Mike Johnson
}
