% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_usgs_gages.R
\name{find_USGS_gages}
\alias{find_USGS_gages}
\title{Locate all USGS NWIS Stream Gages within an Area of Interest}
\usage{
find_USGS_gages(state = NULL, county = NULL, clip_unit = NULL,
  keep.boundary = FALSE, keep.basemap = FALSE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with gage data in a list}

\item{keep.basemap}{logical. If TRUE, the google basemap will be returned with gage data in a list}
}
\description{
Function to locate all USGS stations within and Area of Interest
}
\examples{
Find all USGS stream gages  in Harris County, Texas

harris.usgs = find_USGS_stations(state = 'TX', county = 'harris', keep.boundary = TRUE, keep.basemap = TRUE)
plot(harris.usgs$basemap)
plot(harris.usgs$boundary, add = TRUE, lwd = 5)
plot(harris.usgs$gages, add = TRUE, lwd = 2, col = "darkgreen")

Get gage IDs

IDs = test$gages$feature_id

}
\author{
Mike Johnson
}
