% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_usgs_stations.R
\name{find_USGS_stations}
\alias{find_USGS_stations}
\title{Locate all USGS NWIS Stream Gages within Area of Interest}
\usage{
find_USGS_stations(state = NULL, county = NULL, clip_unit = NULL,
  keep.boundary = FALSE, keep.basemap = TRUE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with station data in a list}
}
\description{
Function to locate all GCHN stations within Area of Interest
}
\author{
Mike Johnson
}
