% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWBD.R
\name{find_WBD}
\alias{find_WBD}
\title{Find all WBD boundaries by location}
\usage{
find_WBD(HUC = NULL, location = NULL, level = 10, subbasins = FALSE,
  keep.basemap = FALSE)
}
\arguments{
\item{HUC}{a character string. Provide a}

\item{location}{a character string. Can be full name or state abbriviation}

\item{level}{specify whay HUC level you are interested in. Options include, 2,4,6,8,10,12.}

\item{subbasins}{logical. If TRUE, a shapefile of nested basin within specified level retuned as list object}

\item{keep.basemap}{logical. If TRUE, a google basemap will be returned with HUC shapefiles}
}
\description{
Function to locate all WBD basins at all nested levels. Input can be a know HUC code or a location within a basin of interest.
The nested leve of interest should be supplied by the level parameter. The subbassins parameter gets all nested levels beneth the levels specification.
}
\examples{

harris.huc = get_WBD(location = "Harris County, Texas", level = 6, subbasins = T, keep.basemap = T)

plot(harris.huc$map)
plot(harris.huc$huc$huc6, add = T, lwd = 6)
plot(harris.huc$huc$huc8, add = T, pch = 16, lwd = 3)
plot(harris.huc$huc$huc10, add = T,  lwd = 1, border = 'red')

}
\author{
Mike Johnson
}
