% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findACIS.R
\name{findACIS}
\alias{findACIS}
\title{Find Meterologic Stations from the Applied Climate Information System}
\usage{
findACIS(AOI, ids = FALSE, param = NULL)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{ids}{If TRUE,  a vector of unique acis station IDs is added to retuned list (default = \code{FALSE})}

\item{param}{what observation type to search for: options include:
\itemize{
\item 'maxt'  :  Maximum temperature (Fahrenheit)
\item 'mint'  :  Maximum temperature (Fahrenheit)
\item 'avgt'  :  Maximum temperature (Fahrenheit)
\item 'prcpn' :  Precipitation       (inches)
\item 'snwd'  :  Snow depth          (inches)
\item 'snow'  :  Snowfall            (inches)
}}
}
\value{
a list() of minimum length 2: AOI and and acis
}
\description{
Query station information from the applied climate information system (ACIS) API. \code{findACIS} returns a \code{SpatailPointsDataframe} of all
stations within an AOI. Some stations are a part of multiple networks. As such each SpatialPoint is marked by all networks and unique station ids within each.
Data comes from the \href{http://www.rcc-acis.org/index.html}{ACIS} and includes the following attributes:
\itemize{
\item 'uid'   : \code{character}  Unique station ID
\item 'NAME'   : \code{character}  Station name
\item 'state'    : \code{integer}    USA state
\item 'elev'    : \code{integer}    Station Elevation (feet)
\item 'staidX'   : \code{character}   Unique station id within network
\item 'network_typeX'    : \code{integer}   Station network
\item 'minDate'    : \code{integer}    Year of first observation
\item 'maxDate'    : \code{integer}    Year of last observation
}\cr
}
\examples{
\dontrun{
sta = getAOI(state = "CO", county = "El Paso") \%>\% findACIS()
}
}
\author{
Mike Johnson
}
