% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findASOS.R
\name{findASOS}
\alias{findASOS}
\title{Find stations from the Automated Surface Observing System}
\usage{
findASOS(AOI, ids = FALSE)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{ids}{If TRUE,  a vector of unique Air Forces Station IDs is added to retuned list (default = \code{FALSE})}
}
\value{
a list() of minimum length 2: AOI and and asos
}
\description{
Query station information from the automated surface observing systems (ASOS) API. \code{findASOS} returns a \code{SpatailPointsDataframe} of all
stations within an AOI. Some stations have data dating back 1901.
All data comes from  \href{http://www.nws.noaa.gov/ost/asostech.html}{NOAA} and includes the following attributes:
\itemize{
\item 'USAF'   : \code{character}  Air Force station ID. May contain a letter in the first position.
\item 'WBAN'   : \code{character}  NCDC WBAN number
\item 'NAME'    : \code{integer}    Station Name
\item 'CTRY'    : \code{integer}    FIPS country ID
\item 'ST'   : \code{character}   State for US stations
\item 'ICAO'    : \code{integer}   ICAO Airport ID
\item 'Start'    : \code{integer}    Beginning Period Of Record (YYYY-MM-DD)
\item 'End'    : \code{integer}    Ending Period Of Record (YYYY-MM-DD)
}\cr
}
\examples{
\dontrun{
sta = getAOI(state = "CO", county = "El Paso") \%>\% findACIS()
}
}
\author{
Mike Johnson
}
