% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCDL.R
\name{findCDL}
\alias{findCDL}
\title{Find Cropland Land Cover Data from the Cropand Data Layer}
\usage{
findCDL(AOI, year = 2017)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{year}{the year(s) to download. Options include 2008-2017 for most locations. Default = 2017}
}
\value{
a list() of minimum length 2: AOI and cld
}
\description{
\code{findCLD} returns \code{Raster} land cover data from the USDA Cropland Data Layer Dataset (\href{https://nassgeodata.gmu.edu/CropScape/}{Cropscape}) for an AOI.
}
\examples{
\dontrun{
 dt = getAOI(clip = list("Devil Tower")) \%>\% findCDL()
 dt = getAOI(clip = list("Devil Tower")) \%>\% findCDL(2011:2016)
}
}
\author{
Mike Johnson
}
