% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAirports.R
\name{findClosestAirports}
\alias{findClosestAirports}
\title{Find Closest Airports}
\usage{
findClosestAirports(location = NULL, n = 5)
}
\arguments{
\item{location}{given as name or a lon, lat pair}

\item{n}{Prescribe how many airports should be returned?}
}
\value{
A table of Airport data including identification codes, names, and distance from prescribed location
}
\description{
\code{findClosestAirports} returns a list and \code{SpatialPointsDataFrame*} of airports closest to a defined location\cr
Returned \code{list} can be interactivly explored via \code{\link{explore}} and ID values (\code{ids = TRUE}) allow for Weather Underground data access via \code{getWU}.\cr\cr
}
\examples{
\dontrun{
# Find Airport nearest you

ap = findClosestAirports(location = findLatLong(), n = 1)

# Find 5 closest Airports to Cal Poly SLO

ap = findClosestAirports(location = 'Cal Poly SLO', n =5)

explore(ap)
}

}
\seealso{
\code{\link{findLatLong}}

\code{\link{findAirports}}

\code{\link{getWU}}

Other HydroData 'find' functions: \code{\link{findAirports}},
  \code{\link{findGHCN}}, \code{\link{findHUC}},
  \code{\link{findNHD}}, \code{\link{findNID}},
  \code{\link{findReservoir}}, \code{\link{findRoads}},
  \code{\link{findSnotel}}, \code{\link{findUSGS}},
  \code{\link{findWaterbodies}}
}
\author{
Mike Johnson
}
