% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGHCN.R
\name{findGHCN}
\alias{findGHCN}
\title{Find NOAA Maintained GHCN Stations}
\usage{
findGHCN(AOI = NULL, param = NULL, ids = FALSE)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{param}{Returned results can be filtered by a requested parameter valid parameters can be found calling \code{unique(ghcn_stations$PARAMETER)}}

\item{ids}{If TRUE,  a vector of station IDs is added to retuned list (default = \code{FALSE})}
}
\value{
a list() of minimum length 2: AOI and ghcn
}
\description{
\code{findGHCN} returns a \code{SpatailPointsDataframe} of all
Global Historical Climatology Network (GHCN) stations within an AOI.
Data comes from the \href{https://www.ncdc.noaa.gov/data-access/land-based-station-data/land-based-datasets/global-historical-climatology-network-ghcn}{NOAA} and includes the following attributes:
\itemize{
\item 'ID'   : \code{character}  Station ID
\item 'NAME'   : \code{character}  Station Name
\item 'LAT': \code{numeric}  Latitude, decimal degrees
\item 'LON'   : \code{numeric} Longitude, decimal degrees
\item 'PARAMETER'   : \code{character}   Observed parameter
\item 'START_YEAR'    : \code{integer}   Year of first observation
\item 'END_YEAR'    : \code{integer}    Year of last observation
}
}
\examples{
\dontrun{
# Get all stations and parameters
sta = getAOI(state = "CO", county = "El Paso") \%>\% findGHCN()

#Limit call to precipiation gages
prcp.sta = getAOI(state = "CO", county = "El Paso") \%>\% findGHCN(param = 'PRCP')

}
}
\author{
Mike Johnson
}
