% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWS.R
\name{findHUC}
\alias{findHUC}
\title{Find Watershed Boundaries (WBD)}
\usage{
findHUC(AOI = NULL, level = 8, subbasins = FALSE, ids = FALSE)
}
\arguments{
\item{level}{defines the HUC level of interest}

\item{subbasins}{If TRUE, all subbasins of the supplied \code{level} will be joined to retuned list}

\item{ids}{If TRUE, returns a list of HUC8 in AOI}

\item{state}{Full name(s) or two character abbriviation(s). Not case senstive}

\item{county}{County name(s). Requires \code{state} input.}

\item{clip}{SpatialObject* or list. For details see \code{getClipUnit}}

\item{save}{If TRUE, data is written to a HydroData folder in users working directory.}

\item{HUC8}{option to supply known HUC8 code(s)}
}
\value{
\code{findHUC} returns a named list of minimum length 1:

\enumerate{
\item 'huc_': A \code{SpatialPolygonDataFrame*} for each requested HUC level defined by \code{level} and \code{subbasins} \cr

Pending parameterization, \code{findWS} can also return:

\item 'ids':       A vector of HUC8 units if \code{ids = TRUE}
}
}
\description{
\code{findWS} returns a list of \code{Spatial*} Objects cropped to an Area of Interest.\cr\cr
To better understand defining an AOI using '\emph{state}', '\emph{county}' and '\emph{clip}' see \code{getAOI} and \code{getClipUnit}.\cr\cr
Returned \code{list} can be interactivly explored via \code{\link{explore}}.\cr\cr
All outputs are projected to \code{CRS '+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs'} and spatial watershed data are (\emph{down})loaded from the  \href{https://prd-tnm.s3.amazonaws.com}{USGS Staged Data Server}.
}
\examples{
\dontrun{
# Find all HUC8, 10 and 12 units surrounding the National Water Center

 ws = findWS(clip = list("National Water Center", 10 ,10),
            subbasins = T,
            level = 8)

# Static Mapping

plot(ws$basemap)
plot(ws$boundary, add = T)
plot(ws$huc_8, lwd = 3, add = T)
plot(ws$huc_10, border = 'red', lwd = 2, add = T)
plot(ws$huc_12, border = 'blue', add = T)

# Generate Interactive Map

explore(ws)
}

}
\seealso{
\code{\link{getAOI}}

\code{\link{explore}}

Other HydroData 'find' functions: \code{\link{findAirports}},
  \code{\link{findGAGESII}}, \code{\link{findGHCN}},
  \code{\link{findNHD}}, \code{\link{findNID}},
  \code{\link{findNWIS}}, \code{\link{findSnotel}},
  \code{\link{findWaterbodies}}
}
\author{
Mike Johnson
}
