% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLatLong.R
\name{findLatLong}
\alias{findLatLong}
\title{Find your Lat, long from IP Address}
\usage{
findLatLong()
}
\value{
A lat, long pair \code{data.frame} Object
}
\description{
Not a good idea for reproducable code but helpful for finding and recording your location
}
\examples{
\dontrun{
# Find your Location

my.location = findLatLong()

# Use with \\emph{clip_unit} calls

usgs = findUSGS(clip_unit = list(findLatLong(), 20, 20))
explore(usgs)

# Use with find closest functions:

ap = findClosestAirports(location = findLatLong(), number = 5)
explore()

}
}
\seealso{
\code{getClipUnit}

\code{getAOI}
}
\author{
Mike Johnson
}
