% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestAirports.R
\name{findNearestAirports}
\alias{findNearestAirports}
\title{Find the Airports nearest a point}
\usage{
findNearestAirports(point = NULL, n = 5, ids = FALSE, bb = FALSE)
}
\arguments{
\item{point}{a point described by lat/long, can be piped from \link[AOI]{geocode}}

\item{n}{the number of Airports to find (default = 5)}

\item{ids}{If TRUE, a vector of airport ICAO IDs is added to retuned list (default = \code{FALSE})}

\item{AOI}{If TRUE, the geometry of the minimum bounding area of the features is added to returned list  (default = \code{FALSE})}
}
\value{
a list() of minimum length 2: AOI and ap
}
\description{
\code{findNearestAirports} returns a \code{SpatialLinesPointsframe} of the 'n' number of airports closest to a declared point.
}
\examples{
\dontrun{
pt = geocode("UCSB") \%>\% findNearestAirports(n = 5)
}
}
\seealso{
findAirports
}
\author{
Mike Johnson
}
