% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestGHCN.R
\name{findNearestGHCN}
\alias{findNearestGHCN}
\title{Find Neareast GHCN stations}
\usage{
findNearestGHCN(location = NULL, n = 5, PARAM = NULL)
}
\arguments{
\item{location}{a lat, long pair}

\item{n}{the number of GHCN to return}

\item{PARAM}{the parameter of interst}
}
\value{
A list contatining a data.frame of Airports and a minimum bounding box data.frame
}
\description{
Find Neareast GHCN stations
}
\author{
Mike Johnson
}
