% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestUSGS.R
\name{findNearestNWIS}
\alias{findNearestNWIS}
\title{Find Neareast USGS stations}
\usage{
findNearestNWIS(point = NULL, n = 5, ids = FALSE, AOI = FALSE)
}
\arguments{
\item{n}{the number of Airports to return}

\item{location}{a lat, long pair or Colloqiual Name}
}
\value{
A list contatining a data.frame of Airports and a minimum bounding box data.frame
}
\description{
Find Neareast USGS stations
}
\author{
Mike Johnson
}
