% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestUSGS.R
\name{findNearestUSGS}
\alias{findNearestUSGS}
\title{Find Neareast USGS stations}
\usage{
findNearestUSGS(location = NULL, n = 5)
}
\arguments{
\item{location}{a lat, long pair or Colloqiual Name}

\item{n}{the number of Airports to return}
}
\value{
A list contatining a data.frame of Airports and a minimum bounding box data.frame
}
\description{
Find Neareast USGS stations
}
\author{
Mike Johnson
}
