% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestWaterbodies.R
\name{findNearestWaterbodies}
\alias{findNearestWaterbodies}
\title{Find Neareast NHD Waterbodies}
\usage{
findNearestWaterbodies(point = NULL, n = 5, ids = FALSE,
  bb = FALSE)
}
\arguments{
\item{point}{a point described by lat/long, can be piped from \link[AOI]{geocode}}

\item{n}{the number of NWIS gages to find (default = 5)}

\item{ids}{If TRUE, a vector of station IDs is added to retuned list (default = \code{FALSE})}

\item{bb}{If TRUE, the geometry of the minimum bounding area of the features is added to returned list  (default = \code{FALSE})}
}
\value{
a list() of minimum length 2: AOI and nwis
}
\description{
\code{findNearestWaterbodies} returns a \code{SpatialPointsDataFrame} of the 'n' number of NHDwaterbodies closest to a declared point.
}
\examples{
\dontrun{
pt = geocode("UCSB") \%>\% findNearestWaterbodies(n = 5)
}
}
\seealso{
findNWIS
}
\author{
Mike Johnson
}
