% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNED.R
\name{findNED}
\alias{findNED}
\title{Find National Elevation Data (NED)}
\usage{
findNED(AOI = NULL, res = 1)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{res}{resolution of NED data. 1 equals 1 arc second, 13 equals 1/3 arc second.}
}
\value{
a list() of minimum length 2: AOI and NED
}
\description{
\code{findNED} returns \code{Raster} elevation data from the National Elevation Dataset (\href{https://nationalmap.gov/elevation.html}{NED}) for an AOI.
Data comes the the USA National Map.
}
\examples{
\dontrun{
 el.paso.elev = getAOI(state = "CO", county = "El Paso") \%>\% findNED(res = 1)
}
}
\author{
Mike Johnson
}
