% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNID.R
\name{findNID}
\alias{findNID}
\title{Find Dams in the US Army Core National Inventory (NID)}
\usage{
findNID(AOI = NULL, ids = FALSE)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{ids}{If TRUE, a vector of Dam IDs is added to retuned list (default = \code{FALSE})}
}
\value{
a list() of minimum length 2: AOI and dams
}
\description{
\code{findNID} returns a \code{SpatialPointsDataFrame}
of all US Army Corps Dams for an Area of Interest from the National Inventory of Dams dataset.
The National Inventory of Dams (NID) is a congressionally authorized database documenting dams in the United States and its territories
This dataset is accessed through the \code{dams} R package and contains 61 attributes, perhaps most notably:
\itemize{
\item 'Dam_Name'   : \code{character}  Dam Name
\item 'NID_ID'   : \code{character}  Unique ID for the dam
\item 'River': \code{character}  Name of the river
\item 'Owner_Type'   : \code{character}  Type of Owner
\item 'Dam_Type'   : \code{character}   Type of Dam
\item 'Primary_Purpose'    : \code{numeric}    Primary Purpose served
\item 'Dam_Length'    : \code{numeric}    Length of the dam
\item 'Dam_Height'   : \code{numeric}  Height of the dam
\item 'Max_Discharge'   : \code{numeric}  Maximum Discharge
\item 'Max_Storage': \code{character}  Maximum Storage
 \item 'Normal_Storage': \code{character}  Normal Storage
} \cr
}
\examples{
\dontrun{
# Find all dams in Texas
tx.dams = getAOI(state = "TX") \%>\% findNID()
}
}
\author{
Mike Johnson
}
