% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNLCD.R
\name{findNLCD}
\alias{findNLCD}
\title{Find National Land Cover Products (NLCD)}
\usage{
findNLCD(AOI = NULL, year = 2011, type = "landcover")
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{year}{the year(s) to download. Options include 2001, 2006, 2011. Default = 2011}

\item{type}{the type of data to downlaod. Options include landcover, canopy, and impervious. Default = landcover}
}
\value{
a list() of minimum length 2: AOI and NLCD
}
\description{
\code{findNLCD} returns \code{Raster} land cover data from the National Land Cover Dataset (\href{https://www.mrlc.gov}{NLCD}) for an AOI.
Data comes the the USA National Map and is avaialble for years 2001, 2006, 2011.
In additon to landcover, users can get data reflecting impervious surface and conaopy cover.
}
\examples{
\dontrun{
 dt = getAOI(clip = list("Devil Tower")) \%>\% findNLDC(2006, 'landcover')
 dt = getAOI(clip = list("Devil Tower")) \%>\% findNLDC(2011, 'canopy')
 dt = getAOI(clip = list("Devil Tower")) \%>\% findNLDC(2011, 'impervious')
}
}
\author{
Mike Johnson
}
