% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNOAAnorms.R
\name{findNOAAnorms}
\alias{findNOAAnorms}
\title{Find stations from NOAAs 1981–2010 US Climate Normals}
\usage{
findNOAAnorms(AOI, ids = FALSE)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{ids}{If TRUE,  a vector of unique acis station IDs is added to retuned list (default = \code{FALSE})}
}
\value{
a list() of minimum length 2: AOI and and acis
}
\description{
Query station information from the Query station information from NOAA 1981-2010 cliamte normats. \code{findNOAAnorms} returns a \code{SpatailPointsDataframe} of all
stations within an AOI. Data comes from the \href{ftp://ftp.ncdc.noaa.gov/pub/data/normals/1981-2010/}{NOAA} and includes the following attributes:
\itemize{
\item 'ID'   : \code{character}  is the station identification code. The first two characters denote the FIPS country code, the third character is a network code identifiNG the station numbering systemused, and the remaining eight characters contain the actual
\item 'ELEV'   : Station name
\item 'ST.ABB'    : USA state abbriviation (USA station only)
\item 'NAME'    : Elevation of the station (in meters, missing = -999.9).
\item 'GSNFLAG'   : ndicates whether the station is part of the GCOS Surface Network (GSN)
\item 'HCNFLAG'    : Indicates whether the station is part of the U.S. Historical Climatology Network (HCN).
\item 'WMOID'    : World Meteorological Organization (WMO) station number
\item 'METHOD'    : Indicates whether a "traditional" or a "pseudonormals" approach was utilized for temperature or precipitation
}\cr
}
\examples{
\dontrun{
sta = getAOI(state = "CO", county = "El Paso") \%>\% findNOAAnorms()
}
}
\author{
Mike Johnson
}
