% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNWIS.R
\name{findNWIS}
\alias{findNWIS}
\title{Find USGS NWIS Stream Gages}
\usage{
findNWIS(AOI = NULL, ids = FALSE, comids = FALSE)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{ids}{If TRUE, a vector of NIWS gage IDs are added to retuned list (default = \code{FALSE})}

\item{comids}{f TRUE, a vector of NHD COMIDs IDs are added to retuned list (default = \code{FALSE})}
}
\value{
a list() of minimum length 2: AOI and nwis
}
\description{
\code{findNWIS} returns a \code{SpatialPointsDataFrame}
of all USGS NWIS gages for an Area of Interest. This dataset is accessed through the NWIS web portal and contains the following attributes:
\itemize{
\item 'OBJECTID'   : \code{character}  Unique ID in dataset
\item 'feature_id'   : \code{character}  NHD COMID for reach
\item 'site_no': \code{character}  NWIS ID number
\item 'site_name'   : \code{character}  Name of site
\item 'da_sqkm'   : \code{character}   Area drainign to gage in square kilometers
\item 'lat_reachCent'    : \code{numeric}    Latitude of the reach center, decimil degrees
\item 'lon_reachCent'    : \code{numeric}    Longitude of the reach center, decimil degrees
}
}
\examples{
\dontrun{
co = getAOI(state = "CO") \%>\% findNWIS()
}
}
\author{
Mike Johnson
}
