% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findReservoir.R
\name{findReservoir}
\alias{findReservoir}
\title{Find USGS recorded USA Reservoirs (Point Data)}
\usage{
findReservoir(state = NULL, county = NULL, clip_unit = NULL,
  boundary = FALSE, basemap = FALSE, save = FALSE, ids = FALSE)
}
\arguments{
\item{state}{Full name(s) or two character abbriviation(s). Not case senstive}

\item{county}{County name(s). Requires \code{state} input.}

\item{clip_unit}{SpatialObject* or list. For details see \code{getClipUnit}}

\item{boundary}{If TRUE, the AOI \code{SpatialPolygon(s)} will be joined to returned list}

\item{basemap}{If TRUE, a basemap will be joined to returned list

 If a user wants greater control over basemap apperance replace TRUE with either:
\itemize{
\item't':  google terrain basemap
\item's':  google sattilite imagery basemap
\item'h':  google hybrid basemap
\item'r':  google roads basemap
}}

\item{save}{If TRUE, data is written to a HydroData folder in users working directory.}

\item{ids}{If TRUE, returns a list of reservoir names}
}
\value{
\code{findReservoir} returns a list of minimum length 1:

\enumerate{
\item 'reservoirs': A \code{SpatialPointsDataFrame}\cr

Pending parameterization, \code{findReservoir} can also return:

\item 'basemap':   A \code{RasterLayer*} basemap if \code{basemap = TRUE}
\item 'boundry':   A \code{SpatialPolygon*} of AOI if \code{boundary = TRUE}
\item 'fiat':      A \code{SpatialPolygon*} of intersected county boundaries if \code{boundary = TRUE}
\item 'ids':       A vector of reservoir names if \code{ids = TRUE}
}
}
\description{
\code{findReservoir} returns a list of \code{Spatial*} Objects cropped to an Area of Interest.\cr\cr
To better understand defining an AOI using '\emph{state}', '\emph{county}' and '\emph{clip_unit}' see \code{getAOI} and \code{getClipUnit}.\cr\cr
Returned \code{list} can be interactivly explored via \code{\link{explore}}.\cr\cr
All outputs are projected to \code{CRS '+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs'} and reservoir data is (\emph{down})loadeded from the \href{https://water.usgs.gov/GIS}{USGS GIS portal}
}
\examples{
\dontrun{
# Find all reservoirs in Texas

tx.res = findReservoir(state = "TX", boundary = TRUE, basemap = TRUE, save = TRUE)

# Static Maping

plot(tx.res$basmap)
plot(tx.res$boundary, add = TRUE)
plot(tx.res$reservoirs, add = TRUE)

# Generate Interactive Map

explore(tx.res)
}

}
\seealso{
\code{\link{getAOI}}

\code{\link{explore}}

Other HydroData 'find' functions: \code{\link{findAirports}},
  \code{\link{findClosestAirports}},
  \code{\link{findGHCN}}, \code{\link{findNHD}},
  \code{\link{findNID}}, \code{\link{findRoads}},
  \code{\link{findSnotel}}, \code{\link{findUSGS}},
  \code{\link{findWS}}, \code{\link{findWaterbodies}}
}
\author{
Mike Johnson
}
