% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findReservoir.R
\name{findReservoirs}
\alias{findReservoirs}
\title{Locate all Reservoirs within Area of Interest}
\usage{
findReservoirs(state = NULL, county = NULL, clip_unit = NULL,
  keep.boundary = FALSE, keep.basemap = FALSE, save = FALSE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with gage data in a list}

\item{keep.basemap}{logical. If TRUE, the google basemap will be returned with gage data in a list}

\item{save}{logical. If TRUE, all data is written to a HydroData folder in the working directory}
}
\description{
Function to locate all US reservoirs within an Area of Interest from the USGS Inventory of reservoirs
}
\examples{
\dontrun{
#Find all reservoirs in Texas

tx.res = find_reservoir(state = "TX", keep.boundary = TRUE, keep.basemap = TRUE, save = TRUE)
plot(tx.res$basmap)
plot(tx.res$boundary, add = TRUE)
plot(tx.res$reservoirs, add = TRUE)
}
}
\author{
Mike Johnson
}
