% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRoads.R
\name{findRoads}
\alias{findRoads}
\title{Find US Census Bureau TIGER Road Networks}
\usage{
findRoads(AOI = FALSE)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}
}
\value{
a list() of minimum length 2: AOI and tiger
}
\description{
\code{findRoads} returns a \code{SpatialLinesDataFrame} of TIGER road networks cropped to an Area of Interest.
This dataset is accessed through the US Census geo portal and contains the following attributes:
\itemize{
\item 'LINEARID'   : \code{character}  Linear feature identifier
\item 'FULLNAME'   : \code{character}  Feature Name
\item 'RTTYP': \code{character}  Route Type Code
\item 'MTFCC'   : \code{character}  Road Classification
\itemize{
\item S1100   Primary Roads
\item R1011   Railroad Feature (Main, Spur, or Yard)
\item R1051   Carline, Streetcar Track, Monorail, Other Mass Transit Rail)
\item R1052   Cog Rail Line, Incline Rail Line, Tram
\item S1100   Primary Road
\item S1200   Secondary Road
}
}
}
\examples{
\dontrun{
AOI  = getAOI(clip = list("UCSB", 10, 10)) \%>\% findRoads()
}
}
\author{
Mike Johnson
}
