% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSnowtel.R
\name{findsnotel}
\alias{findsnotel}
\title{Locate all NRCS snotel stations within an Area of Interest}
\usage{
findsnotel(state = NULL, county = NULL, clip_unit = NULL,
  keep.boundary = FALSE, keep.basemap = FALSE, ids = FALSE,
  save = FALSE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with gage data in a list}

\item{keep.basemap}{logical. If TRUE, the google basemap will be returned with gage data in a list}

\item{ids}{logical. If TRUE, returns a vector of station IDs}

\item{save}{logical. If TRUE, all data is written to a HydroData folder in the working directory}
}
\description{
Function to locate all snotel stations within and Area of Interest
}
\examples{
\dontrun{
#Find all snotel stations in Nevada

nv.snow = find_snotel(state = 'NV', keep.boundary = TRUE, keep.basemap = TRUE, save = TRUE)
plot(nv.snow$basemap)
plot(nv.snow$boundary, add = TRUE, lwd = 5)
plot(nv.snow$snotel, add = TRUE, pch = 8, col = "lightblue")
}
}
\author{
Mike Johnson
}
