% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWaterbodies.R
\name{findWaterbodies}
\alias{findWaterbodies}
\title{Find Water Bodies within Area of Interest}
\usage{
findWaterbodies(state = NULL, county = NULL, clip_unit = NULL,
  boundary = FALSE, basemap = FALSE, save = FALSE)
}
\arguments{
\item{state}{character. Provide full name(s) or two character abbriviation(s). Not case senstive}

\item{county}{character. Provide county name(s). Requires \code{state} input.}

\item{clip_unit}{SpatialObject* or list. For details see \code{?getClipUnit}}

\item{boundary}{logical. If TRUE, the AOI \code{SpatialPolygon(s)} will be joined to returned list}

\item{basemap}{logical. If TRUE, a basemap will be joined to returned list from \code{\link[dismo]{gmap}}.

 #' If a user wants greater control over basemap apperance replace TRUE with either:
\itemize{
\item't':  a terrain imagery basemap
\item's':  a sattilite imagery basemap
\item'h':  a hybrid imagery basemap
\item'r':  a roadmap imagery basemap
}}

\item{save}{logical. If TRUE, all data is saved to a HydroData folder created in users working directory. Find working directory with \code{\link[getwd()]}}
}
\description{
\code{findWaterbodies} finds polygon waterbodies within an Area of Interest from the ESRI Global Waterbodies database
}
\details{
\code{findWaterbodies} returns a named list of minimum length 1:
\enumerate{
\item 'wb': A \code{SpatialLinesDataFrame} of NHD flowlines and metadata
\item 'basemap':   A \code{RasterLayer} basemap if 'basemap' is \code{TRUE}
\item 'fiat':      A \code{SpatialPolygon} of fiat boundaries if 'boundary' is \code{TRUE}
\item 'clip':      A \code{SpatialPolygon} of clip unit boundary if 'boundary' is \code{TRUE}
\item 'ids':       A vector of flowline COMIDs if 'ids' is \code{TRUE}
}
}
\examples{
\dontrun{
#Find all water bodies in San Luis Obispo, CA

slo.wb = findWaterbodies(state = 'CA', county = "San Luis Obispo",
                         boundary = TRUE, basemap = TRUE)

plot(slo.wb$basemap)
plot(slo.wb$boundary, add = T)
plot(slo.wb$waterbodies, add = T, col = 'blue')
}
}
\author{
Mike Johnson
}
