% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWaterPoly.R
\name{findWaterPoly}
\alias{findWaterPoly}
\title{Find Water Polygon}
\usage{
findWaterPoly(AOI)
}
\arguments{
\item{AOI}{an AOI}
}
\value{
a single polygon of all NHD water features.
}
\description{
A function to join all NHD water featrues into a single polygon.
Flowlines are sized according to there Strahler stream order multiplied by 10 m.
Useful for viualization and masking operations where a water surface is needed.
}
\examples{
\dontrun{
getAOI(state = 'CO', county = 'El Paso') \%>\% findWaterPoly()
}
}
\author{
Mike Johnson
}
