% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWS.R
\name{findWBD}
\alias{findWBD}
\title{Find Watershed Boundary Geometries (WBD/HUC)}
\usage{
findWBD(AOI = NULL, level = 8, subbasins = FALSE, crop = TRUE,
  ids = FALSE)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{level}{defines the HUC level of interest (default = 8)}

\item{subbasins}{If TRUE, all subbasins of the supplied level will be joined to retuned list}

\item{crop}{If TRUE, all objects are cropped to the AOI boundaries (default = \code{TRUE})}

\item{ids}{If TRUE,  a vector of finest resolution HUC codes is added to retuned list (default = \code{FALSE})}
}
\description{
The United States Geological Survey maintains a hierarchical system of hydrologic units each assigned a
unique code (HUC). The heirarchical level is described by the number of digits in the code. A two-diget code (eg. HUC 2) is the coarsest unit of aggregation while the HUC 12 is the finest resulution.
The spatail geometries of these units are stored in the Watershed Boundary Dataset with coverage of the United States.
\code{findWBD} returns a \code{SpatialPolygonsDataFrame*} of WBD boundaries for the specified level within an AOI. Pending the query,
data comes from the USGS CIDA server or the USGS staged products FTP.\cr

Below you can see the general factors for each HUC level:\cr
\tabular{ccccc}{
  Name \tab Digits \tab Average Size (sqmiles) \tab Example Name \tab Example Code \cr
 Region \tab 2 \tab 177,560 \tab Pacific Northwest \tab 17 \cr
Subregion \tab 4 \tab  16,800  \tab Lower Snake \tab 1706\cr
 Basin \tab 6 \tab 10,596  \tab Lower Snake \tab 170601 \cr
  Subbasin \tab 8 \tab 700  \tab Imnaha River \tab 17060102 \cr
  Watershed \tab 10 \tab 227 \tab Upper Imnaha River \tab 1706010201 \cr
  Subwatershed \tab 12 \tab 40 \tab North Fork Imnaha River \tab 170601020101 \cr
}
}
\examples{
\dontrun{
# Get Cropped HUC8s for AOI
 getAOI(clip = list("UCSB", 10, 10)) \%>\% findWBD()

# Get Cropped HUC10s for AOI
 getAOI(clip = list("UCSB", 10, 10)) \%>\% findWBD(level = 10)

# Get Cropped HUC8s, HUC10s and HUC12s for AOI
 getAOI(clip = list("UCSB", 10, 10)) \%>\% findWBD(level = 8, subbasins = TRUE)

# Get uncropped HUC10s for AOI
 getAOI(clip = list("UCSB", 10, 10)) \%>\% findWBD(level = 10, crop = FALSE)
}
}
\author{
Mike Johnson
}
