% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWS.R
\name{findWS}
\alias{findWS}
\title{Find Watershed boundaries in AOI}
\usage{
findWS(state = NULL, county = NULL, clip_unit = NULL, level = 10,
  subbasins = FALSE, HUC8 = NULL)
}
\arguments{
\item{HUC8}{}
}
\description{
Find Watershed boundaries in AOI
}
\examples{

}
