% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.rating.R
\name{get.rating}
\alias{get.rating}
\title{Find Synthetic Rating Curve Information}
\usage{
get.rating(flowlines)
}
\arguments{
\item{flowlines.}{A shapefile of NHD flowlines.}
}
\description{
Download and subset a synthetic rating curves for a set of NHD flowlines
}
\examples{
al.rating = get.rating(flowlines = al.flows)

Allows for internal nesting of functions:
This examples nests get_ip_loc(), find_flowlines(), and get.rating()

ucsb.rating = get.rating(flowlines = find_flowlines(clip_unit = list(get_ip_loc(), 10, 10)))

Or for a watershed

ucsb.rating = get.rating(flowlines = find_flowlines(clip_unit = get_WBD(location = "UCSB", level =10)))

}
\author{
Mike Johnson
}
