% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HAND.R
\name{get_HAND}
\alias{get_HAND}
\title{Get HUC6 Catchment Boundaries, and HAND Products}
\usage{
get_HAND(state = NULL, county = NULL, clip_unit = NULL,
  keep.boundary = FALSE, keep.flowlines = TRUE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with gage data in a list}
}
\value{
This function downloads the associated shapefiles into the /Geospatial/Flowlines folder built using
\code{\link{build_files}}
}
\description{
Downloads and subsets Height Above Nearest Drainage (HAND) and Catchment raster data from the CyberGIS FTP server for an area of interest.
HUC6 HAND raster are very large, so inital processing can be time consuming
}
\examples{
al.hand = get_HAND(clip_unit = list("National Water Center", 10, 10))
plot(al.hand$hand, col = rev(blues9))
plot

}
\seealso{
\code{\link{build_files}}
\code{\link{get_rating_curve}}
}
\author{
Mike Johnson
}
