% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ip_loc.R
\name{get_ip_loc}
\alias{get_ip_loc}
\title{A function for returning a users Lat, long from IP Address}
\usage{
get_ip_loc()
}
\value{
A numeric lat, long pair
}
\description{
A function for returning a users Lat, long from IP Address
}
\examples{
\dontrun{
my.location = get_loc_ip()
}
}
\author{
Mike Johnson
}
