% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stage.R
\name{get_stage}
\alias{get_stage}
\title{Convert Discharge to Stage via HAND synthetic rating curves}
\usage{
get_stage(flows, rating_curves)
}
\arguments{
\item{flows}{dataframe of flows generated from readNWM \code{\link{readNWM}}}

\item{rating_curves}{dataframe of synthetic rating curves \code{\link{get.rating}}}
}
\description{
Converts data, which can be found via \code{\link{readNWM}} to stage values
using the sythetic rating curves assoicated with the HAND methodology. These rating curves can be generated via
the \code{\link{get.rating}}.
}
\author{
Mike Johnson
}
