% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundaries.R
\name{getBasemap}
\alias{getBasemap}
\title{Get a Google base Map via Dismo package gmap}
\usage{
getBasemap(AOI = NULL, type = NULL)
}
\arguments{
\item{AOI}{an Area of Intrest Object generated with \code{getAOI}}

\item{type}{detemrines the type of base map returned. Choose from:
\itemize{
\item 't':  a terrain imagery basemap (DEFAULT IF  \code{NULL})
\item 's':  a sattilite imagery basemap
\item 'h':  a hybrid imagery basemap
\item 'r':  a roadmap imagery basemap
}}
}
\description{
\code{getBasemape} wraps the \code{\link[dismo]{gmap}} function in the framework of HydroData
}
\examples{
\dontrun{
 plot(getBasemap(getAOI(state = "CA", county = "Santa Barbara"), type = 's'))

}

}
\seealso{
\code{\link[dismo]{gmap}}
}
\author{
Mike Johnson
}
\concept{HydroData 'get' function}
