% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_get_boundingbox.R
\name{getBoundingBox}
\alias{getBoundingBox}
\title{Get Bounding box}
\usage{
getBoundingBox(x)
}
\arguments{
\item{x}{a \code{data.frame} with a lat and long column}
}
\value{
a \code{SpatialPolygon} bounding box of input points \code{x}
}
\description{
A function to define a minimum bounding box for a set of points
}
\seealso{
Other HydroData 'utility' function: \code{\link{download_shp}}
}
