% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDaymet_pixel.R
\name{getDaymet_pixel}
\alias{getDaymet_pixel}
\title{Get Pixel level Daymet Data for a lat long pair}
\usage{
getDaymet_pixel(lat = NULL, lon = NULL, start = NULL, end = start + 1)
}
\arguments{
\item{lat}{(numeric) a latitude location}

\item{lon}{(numeric) a longitude location}

\item{start}{(numeric) the first year to grab data}

\item{end}{(numeric) the end year to grab data to}
}
\value{
A HydroData object with lat, lon, elevation, tileID, and data table
}
\description{
\code{getDaymet_pixel} returns a HydroData object and \code{SpatialPointsDataFrame*} of airports closest to a defined location\cr
Returned \code{list} can be interactivly explored via \code{\link{inspect}}.\cr\cr
}
\details{
Parmeters returned include:
\itemize{
 \item{"year"}{ The year}
 \item{"yday"}{ The Julien Day}
 \item{"dayl"}{ Duration of the daylight period for the day in seconds}
 \item{"prcp"}{ Daily total precipitation in mm}
 \item{"srad"}{ Avergae Incident shortwave radiation flux density in w/m2}
 \item{"swe"}{ Snow Water Equivilent in kg/m2 }
 \item{"tmax"}{ Daily maximum 2-meter air temperature in degrees C}
 \item{"tmin"}{ Daily minimum 2-meter air temperature in degrees C}
 \item{"vp"}{ Water Vapor Pressure in Pa}
 \item{"Date"}{ Date}
 \item{"agency_cd"}{ DAYMET}
 \item{"site_no"}{ 1}
}
}
\examples{
\dontrun{
 # Get all Daymet Data for UCSB from 2015 - 2016
 ucsb.daymet =  getDaymet_pixel( lat = 34.4208,
                                 lon = -119.6982,
                                 start = 2015,
                                 end = 2016)

 inspect(data = ucsb.daymet$data, param = "srad")
}

}
\seealso{
\code{\link{getGHCN}}

\code{\link{getDaymet}}

\code{\link{inspect}}

Other HydroData 'get' functions: \code{\link{getAOI}},
  \code{\link{getClipUnit}}, \code{\link{getFiatBoundary}},
  \code{\link{getUSGS}}
}
\author{
Mike Johnson
}
