% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiatBoundary.R
\name{getFiatBoundary}
\alias{getFiatBoundary}
\title{A function for finding County and State shapefiles}
\usage{
getFiatBoundary(state = NULL, county = NULL)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string.}
}
\value{
SpatialPolygonDataFrame
}
\description{
Generates a shapefile for a county or state from a user query complete with FIP code, and identifying names.
}
\examples{
get_region_sp(state = "CA")
get_region_sp(state = c("CA","Utah","Nevada"))
get_region_sp(state = "CA", county = "San Luis Obispo")
get_region_sp(state = "CA", county = c("San Luis Obispo", "Santa Barbara", "Ventura")

}
\author{
Mike Johnson
}
