% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGCHN.R
\name{getGCHN}
\alias{getGCHN}
\title{Get GCHN data from list of station IDs}
\usage{
getGCHN(IDs = NULL, parameters = NULL, years = NULL)
}
\arguments{
\item{IDs}{a list or single of station IDs}

\item{parameters}{a character string. Should the returned data be subset to a particular parameter(s)?}

\item{years}{a character or numeric string. Should the returned data be subset to a particular year(s)?}
}
\description{
Function to download a dataframe of meterological data from known GCHN station. See findGCHN function to find station ids by AOI.
}
\examples{

}
\author{
Mike Johnson
}
