% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGCHN.R
\name{getGHCN}
\alias{getGHCN}
\title{Get GHCN data from list of station IDs}
\usage{
getGHCN(IDs = NULL, parameters = NULL, years = NULL)
}
\arguments{
\item{IDs}{a list or single of station IDs}

\item{parameters}{a character string. Should the returned data be subset to a particular parameter(s)?}

\item{years}{a character or numeric string. Should the returned data be subset to a particular year(s)?}
}
\description{
Function to download a dataframe of meterological data from known GHCN station. See findGHCN function to find station ids by AOI.
}
\author{
Mike Johnson
}
