% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKoppenClass.R
\name{getKoppenClass}
\alias{getKoppenClass}
\title{Gets and summarizes Koppen Climate Classification Data for AOI}
\usage{
getKoppenClass(state = NULL, county = NULL, clip_unit = NULL,
  keep.basemap = FALSE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or vector defining centroid and bounding box height and width in miles}

\item{keep.basemap}{logical. If TRUE, the google basemap will be returned with gage data in a list}
}
\value{
Raster
SpatialPointsDataFrame
Summary Dataframe
}
\description{
Generates a raster, spatial point and summary of Koppen Classifications
}
\examples{
\dontrun{
co.koppen = getKoppenClass(state = "CA")
}
}
\author{
Mike Johnson
}
