% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNED.R
\name{getNED}
\alias{getNED}
\title{Get NED data for an Area of Interest}
\usage{
getNED(state = NULL, county = NULL, clip_unit = NULL, res = 1,
  keep.boundary = FALSE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{res}{resolution of NED data. 1 equals 1 arc second, 13 equals 1/3 arc second.}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with station data in a list}
}
\description{
Function to download elevation data for an area of interest
}
\examples{
\dontrun{
#Get 1 arc second elevation for El Paso County Colorado Springs

el.paso.elev = get_ned(state = "CO", county = "El Paso", res = 1, keep.boundary = TRUE)
plot(el.paso.elev, col = col_elev)
}
}
\author{
Mike Johnson
}
