% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNLDI.R
\name{getNLDI}
\alias{getNLDI}
\title{Get NLDI Data}
\usage{
getNLDI(comid = NULL, nwis = NULL, huc12 = NULL, type = NULL,
  spatial = TRUE, resources = NULL)
}
\arguments{
\item{comid}{a COMID value (see \code{findNHD} and \code{findNearestCOMID})}

\item{nwis}{a USGS station values (see \code{findUSGS} and \code{findNearestUSGS})}

\item{huc12}{a HUC12 id (see \code{findHUC})}

\item{type}{a type of data to return ("UT", "UM","DD", "DM", "basin")}

\item{spatial}{if \code{TRUE} returned data will be sp, else of type sf}

\item{resources}{what type of resources to add ("comid", "huc12pp", "npdes_rad", "nwis", "wqp")}
}
\value{
a list of spatial / sf features
}
\description{
This function acts as a formal query to the NLDI
}
\author{
Mike Johnson
}
