% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUSGS.R
\name{getUSGS}
\alias{getUSGS}
\title{Get USGS data}
\usage{
getUSGS(IDs = NULL, save = FALSE, parameter = "00060")
}
\arguments{
\item{IDs}{NWIS station ID(s)}

\item{save}{If TRUE, data is written as a csv to a HydroData folder in users working directory.}

\item{paramater}{Default is streamflow (00060), but can be altered}
}
\value{
\code{getUSGS} returns a \code{data.frame*} of observation data
}
\description{
\code{findUSGS} returns a list of \code{data.frame*} Object for stations of interst.\cr\cr

THIS IS A SHAMELESS WRAPPER AROUND THE USGS dataRetrival \code{readNWISdv} function TO FIT THE HYDRODATA FRAMEWORK !!! \cr\cr

By default this function will return streamflow data for a list of stations found with \code{getUSGS}
If you want a parameter other then steamflow, see list of parameters \href{https://help.waterdata.usgs.gov/codes-and-parameters/parameters}{here}\cr
Returned \code{data.frame} can be interactivly explored via \code{\link{inspect}}
}
\examples{
\dontrun{
# Find USGS station near UCSB

 sta = findUSGS(clip_unit = list("UCSB", 10, 10), ids = T)

# Generate Interactive Map

 explore(sta)

# Get histroic streamflow data for those stations

 data = getUSGS(IDs = sta$ids, save = T)

# View interactive timeseries of daily, and yearly averages

 inspect(data, param = 'Flow', timestep = "daily")
 inspect(data, param = 'Flow', timestep = "yearly")

}

}
\seealso{
\code{\link{findUSGS}}

\code{\link{inspect}}

Other HydroData 'get' functions: \code{\link{getAOI}},
  \code{\link{getClipUnit}}, \code{\link{getFiatBoundary}}
}
\author{
Mike Johnson
}
\concept{HydroData 'get' functions}
