% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{ghcn_stations}
\alias{ghcn_stations}
\title{Global Historical Climatology Network (GHCN) daily data}
\format{a \code{dataframe} instance, 1 row per station with columns:
\itemize{
\item 'ID':         A \code{character} Station ID
\item 'NAME':       A \code{character} Station name
\item 'LAT':        A \code{numeric}   Station latitude
\item 'LON':        A \code{numeric}   Station longitude
\item 'PARAMETER':  A \code{character} Parameter being recorded
\item 'START_YEAR': A \code{integer}   Latitude of NHD reach center
\item 'END_YEAR':   A \code{integer}   Longitude of NHD reach center
}}
\source{
\href{
https://www.ncdc.noaa.gov/data-access/land-based-station-data/land-based-datasets/global-historical-climatology-network-ghcn}{
NOAA National Centers for Environmental Information}
}
\usage{
ghcn_stations
}
\description{
Dataset containing the GCHN daily station network
}
\examples{
\dontrun{
 gchn = HydroData::ghcn_stations
}
}
\keyword{datasets}
