% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroData.R
\docType{package}
\name{HydroData}
\alias{HydroData}
\alias{HydroData-package}
\title{HydroData}
\description{
\code{HydroData} package
}
\details{
Access earth systems geospatial and observation data programatically

See the README on
}
