% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Inspect Timeseries data found with HydroData}
\usage{
inspect(data = NULL, timestep = "daily", param = NULL, save = FALSE)
}
\arguments{
\item{save}{}
}
\description{
\code{inspect} can be used to explore timeseries HydroData objects
Namely:
 getUSGS
 getWU
 getNWM
 getGHCN
}
