% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{kopRas}
\alias{kopRas}
\title{Koppen Climate Classification Raster, March 2017}
\format{a \code{RasterLayer}}
\source{
\href{http://koeppen-geiger.vu-wien.ac.at/present.htm}{Koppen Climate Data}
}
\usage{
kopRas
}
\description{
\code{kopRas} contains the March 2017 Global Koppen Climate Classification Dataset
}
\examples{
\dontrun{
 kop = HydroData::kopRas
}
}
\references{
Kottek, M., J. Grieser, C. Beck, B. Rudolf, and F. Rubel, 2006:
World Map of the Köppen-Geiger climate classification updated. Meteorol. Z., 15, 259-263.
DOI: 10.1127/0941-2948/2006/0130.
}
\keyword{datasets}
