% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.data.R
\name{meta.data}
\alias{meta.data}
\title{Return HydroData Object metadata}
\usage{
meta.data(AOI)
}
\arguments{
\item{AOI}{and AOI object with appended HydroData features}
}
\value{
a data.frame of meta.data describing a HydroData object
}
\description{
Function for generating descriptive meta data for all HyrdroData objects
}
\author{
Mike Johnson
}
