% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\docType{data}
\name{meta}
\alias{meta}
\title{Appliced Climate Information Systems Meta Data}
\format{a \code{list}}
\source{
\href{http://www.rcc-acis.org}{Applied Climate Information System}
}
\usage{
meta
}
\description{
\code{ACIS_meta} contains the meta data needed to query stations from the ACIS API tile index for DAYMET meterological Data
}
\examples{
\dontrun{
 meta = HydroData::meta
}
}
\keyword{datasets}
