% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitiy_mosaic.R
\name{mosaic.hd}
\alias{mosaic.hd}
\title{Mosaic}
\usage{
mosaic.hd(input)
}
\arguments{
\item{input}{raster stack}
}
\description{
Interal function for mosaicing a list of rasters
}
\author{
Mike Johnson
}
