% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_clip_unit.R
\name{name.clip.unit}
\alias{name.clip.unit}
\title{Interprete the plain-english description of a clip unit}
\usage{
name.clip.unit(clip_unit)
}
\arguments{
\item{clip_unit}{a user supplied clip_unit}
}
\value{
a string describing the clipt unit in plain english
}
\description{
Interprete the plain-english description of a clip unit
}
\seealso{
Other HydroData 'helper' functions: \code{\link{define.clip.unit}},
  \code{\link{return.what}}
}
\author{
Mike Johnson
}
