% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAOI.R
\name{nameAOI}
\alias{nameAOI}
\title{A function for returning a text string name of AOI}
\usage{
nameAOI(state = NULL, county = NULL, clip_unit = NULL)
}
\value{
A character string
}
\description{
A function for returning a text string name of AOI
}
\examples{
nameAOI(state = "CA")

}
\author{
Mike Johnson
}
