% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\docType{data}
\name{nhd_outlet}
\alias{nhd_outlet}
\title{NHD COMIDs outlets}
\format{a \code{dataframe} instance, 1 row per station with columns:
\itemize{
\item 'COMID':         A \code{integer} Station COMID
\item 'LAT':        A \code{numeric}   Station latitude
\item 'LON':        A \code{numeric}   Station longitude
}}
\source{
\href{
https://www.ncdc.noaa.gov/data-access/land-based-station-data/land-based-datasets/global-historical-climatology-network-ghcn}{
NOAA National Centers for Environmental Information}
}
\usage{
nhd_outlet
}
\description{
Dataset containing the NHD outlets
}
\examples{
\dontrun{
 gchn = HydroData::comids
}
}
\keyword{datasets}
