% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterUse.R
\name{plotWS}
\alias{plotWS}
\title{Plot a Water use Object}
\usage{
plotWS(waterUse.obj = NULL, variable = NULL, plot.type = "static",
  year = NULL)
}
\arguments{
\item{waterUse.obj}{a water use object from the waterUse() function}

\item{variable}{the water use variable to plot. Can be indexed from wu.variable$...}

\item{plot.type}{can be either a spplot static image of a leaflet derived interactive image. Options include "static" or "interactive"}

\item{year}{the year of data to plot}
}
\description{
Function to visualize waterUse data by county in both static and interactive formats.
}
\examples{
\dontrun{
plotWu(wu10,
       variable = wu.variable$`Public Supply, surface-water withdrawals, fresh, in Mgal/d`,
       plot.type = 'static', year = 2010)

plotWu(wu10,
       variable = wu.variable$`Public Supply, surface-water withdrawals, fresh, in Mgal/d`,
       plot.type = 'interactive', year = 2010)
}
}
\author{
Mike Johnson
}
