% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penman_montieth.R
\name{pm}
\alias{pm}
\title{penman_montieth}
\usage{
pm(Tair = NULL, vpd = NULL, Rnet = NULL, gs = 0.5, ga = NULL,
  CP = 1010, Pair = 101325, dayl = 24)
}
\arguments{
\item{Tair}{(deg C) air temperature}

\item{vpd}{(Pa)    vapour pressure deficit}

\item{Rnet}{(j/m2/day)      net radiation}

\item{gs}{(mm/s)  surface conductance}

\item{ga}{(mm/s)  aerodynamic conductance}

\item{CP}{1010.0 (J/kg*K) specific heat of air}

\item{Pair}{air pressure (default 101325 (Pa)}

\item{dayl}{number of hours in the day (default 24)}
}
\value{
Evapotranspiration (mm/day)
}
\description{
Computes evapotranspiration
}
\author{
Mike Johnson
}
