% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNWM.R
\name{readNWM}
\alias{readNWM}
\title{Hourly Value NWM Data Retrival}
\usage{
readNWM(comids, startDate, endDate = NULL, config = "analysis_assim",
  time = NULL, forecast = NULL, keep.flowlines = FALSE,
  keep.basemap = FALSE, path = NULL, interval.hr = NULL)
}
\arguments{
\item{comids}{used to identify NHD reaches to subset CONUS forecast. Can be given as a numeric vector, a AOI shapefile (use defineAOI()) or a shapefile of NHD flowlines}

\item{startDate}{character starting date for data retrieval in the form YYYY-MM-DD for data retrivial, YYYY-MM-DD-HH-MM-SS for folder of NETCDF files.}

\item{endDate}{character ending date for data retrieval in the form YYYY-MM-DD. Only used with config = 'analysis_assim'}

\item{config}{character describing what forecast congiguration to access. Options are 'analysis_assim', 'short_range', 'medium_range', or 'long_range'}

\item{time}{numeric decribing time a forecast is made in UTC timezone. Must be between 0 and 23.}

\item{forecast}{numeric describing number of hours forcasted out from defined 'time'}

\item{keep.flowlines}{logical. If TRUE, the NHD shapefile will be returned with flow data (cannot be used if comids defined by vector)}

\item{keep.basemap}{logical. If TRUE, the google basemap will be returned with flow data in a list}

\item{path}{character to folder of NETCDF files.}

\item{interval.hr}{logical. Hourly step between folder of NETCDF files}
}
\description{
Imports National Water Model streamflow data from the Hydroshare Data Explore or a folder of previously downloaded NETCDF files.
}
\author{
Mike Johnson
}
