% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_save.file.R
\name{save.file}
\alias{save.file}
\title{Save File}
\usage{
save.file(data = NULL, state = NULL, county = NULL, clip_unit = NULL,
  agency = NULL, source = NULL, dataset = NULL, other = NULL)
}
\arguments{
\item{data}{What data to write}

\item{state}{State abb}

\item{county}{country call}

\item{clip_unit}{clip unit call}

\item{agency}{what agency maintains the data}

\item{source}{what is the source of the data}

\item{dataset}{what is the dataset that is subset}

\item{other}{anything else that should be documented (eg year)}
}
\description{
Save File
}
