% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_save.file.R
\name{saveHD}
\alias{saveHD}
\title{Save File}
\usage{
saveHD(AOI = NULL, path = NULL, ext = "shp")
}
\arguments{
\item{AOI}{What data to write}

\item{path}{State abb}

\item{ext}{country call}
}
\description{
Save File
}
