% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAxis.R
\name{setAxis}
\alias{setAxis}
\title{set Axis}
\usage{
setAxis(data, dates, param, timestep)
}
\arguments{
\item{data}{data input}

\item{dates}{dates present in data}

\item{param}{parameter to name}

\item{timestep}{timestep to aggregate to}
}
\description{
set Axis
}
