% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_data.R
\docType{data}
\name{snotel}
\alias{snotel}
\title{Snotel Stations}
\format{a \code{dataframe} instance, 1 row per station with columns:
\itemize{
\item 'NETWORK':    A \code{integer} Network of interest
\item 'STATE':      A \code{integer}  State abbriviation
\item 'NAME':       A \code{character} Site name
\item 'START.DATE': A \code{integer}    Day of first measurement
\item 'LAT':        A \code{integer} Station Latitude
\item 'LONG':       A \code{character}  Station Longitude
\item 'ELEV':       A \code{character} Station Elevation
\item 'COUNTY':     A \code{character}  County
\item 'HUC12.NAME': A \code{character}  HUC 12 Name
\item 'HUC12.ID':   A \code{character}  HUC 12 code
}}
\source{
\href{https://wcc.sc.egov.usda.gov/nwcc/yearcount?network=sntl&state=&counttype=statelist}{SNOTEL INFORMATION}
}
\usage{
snotel
}
\description{
\code{snotel}NRCS SNOTEL station metadata
}
\examples{
\dontrun{
 tiles = HydroData::snotelStations.rda
}
}
\keyword{datasets}
