% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sf.R
\name{to_sf}
\alias{to_sf}
\title{Convert Spatial HydroData Objects to Simple Features}
\usage{
to_sf(hydro.data.object = NULL)
}
\arguments{
\item{hydro.data.object}{a HydroData object with Spatial components}
}
\value{
a list with the same length as the input
}
\description{
A function to convert all Spatial* HydroData objects to simple feature geometries.
Non-Spatial objects (eg raster and list components) will be skipped over.
}
\examples{
\dontrun{
AOI = getAOI(clip = 'UCSB') \%>\% findNED \%>\% findNHD \%>\% findNWIS \%>\% to_sf
}
}
\author{
Mike Johnson
}
