% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterUse.R
\name{waterUse}
\alias{waterUse}
\title{Get Multi-Year USGS Water Use data}
\usage{
waterUse(state = NULL, county = NULL, clip_unit = NULL, year = 2010)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{year}{define year to download. Any combination of 2000, 2005 and 2010 accepted}
}
\value{
list of SpatialPolygons and Data
}
\description{
Function to download national water use data for an area of interest from 2000, 2005, 2010
}
\examples{
\dontrun{
wu10 = waterUse(state = "CO", year = 2010)

wu.co = waterUse(statte = c("CA", "UT", "AZ", "NV"), year = c(2000, 2005, 2010))
}
}
\author{
Mike Johnson
}
