% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipMerge.R
\name{clipMerge}
\alias{clipMerge}
\title{Clip and Merge raster data}
\usage{
clipMerge(dataset, bounds)
}
\arguments{
\item{dataset}{a raster stack or brick}

\item{bounds}{a clipping boundary}
}
\description{
Clip and Merge raster data
}
\details{
\code{clipMerge} clip raster brick(s) to a set boundary and merge
}
\seealso{
Other HydroData 'helper' function: \code{\link{download.shp}}
}
\author{
Mike Johnson
}
\concept{HydroData 'helper' function}
