% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.shp.R
\name{download.shp}
\alias{download.shp}
\title{Download, unzip and read shapefile from URL}
\usage{
download.shp(URL, type)
}
\arguments{
\item{URL}{path to data}

\item{type}{description of data being downloaded used for messaging}
}
\description{
Download, unzip and read shapefile from URL
}
\examples{
\dontrun{
download.shp(URL, type = "flowlines")
}

}
\seealso{
Other HydroData 'helper' function: \code{\link{clipMerge}}
}
\author{
Mike Johnson
}
\concept{HydroData 'helper' function}
