% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_download_url.R
\name{download.url}
\alias{download.url}
\title{Download data from URL via httr::GET}
\usage{
download.url(url)
}
\arguments{
\item{url}{url}
}
\value{
a status and path to downloaded file
}
\description{
Download file from URL using httr:GET.
File is written to the tmp directory with the basename of the URL call.
}
\author{
Mike Johnson
}
