% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ghcnd_stations.R
\name{find_ghcnd_stations}
\alias{find_ghcnd_stations}
\title{Locate all NOAA GCHN Stations within Area of Interest}
\usage{
find_ghcnd_stations(state = NULL, county = NULL, clip_unit = NULL,
  keep.boundary = TRUE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with station data in a list}
}
\description{
Function to locate all GCHN stations within Area of Interest
}
\examples{
Find all GCHN stations in Harris County, Texas

harris.clim = find_ghcnd_stations(state = "TX", county = "Harris")
plot(harris.clim$boundary)
plot(harris.clim$gchn, add = TRUE, pch = 16, col = 'blue')

}
\author{
Mike Johnson
}
