% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAirports.R
\name{findAirports}
\alias{findAirports}
\title{Find Airport Data}
\usage{
findAirports(AOI = NULL, ids = FALSE)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{ids}{If TRUE,  a vector of airport ICAO codes is added to retuned list (default = \code{FALSE})}
}
\value{
a list of minimum length 2: AOI and ap
}
\description{
\code{findAirports} returns a \code{SpatialPointsDataframe} of all airports within an AOI.
Data comes from the \href{https://openflights.org/data.html}{Openflights} database and includes the following attributes:
\itemize{
\item 'name'   : \code{character}  Name of airport. May or may not contain the City name.
\item 'city'   : \code{character}  Main city served by airport. May be spelled differently from Name.
\item 'country': \code{character}  Country or territory where airport is located.
\item 'IATA'   : \code{character}  3-letter IATA code
\item 'ICAO'   : \code{numeric}    4-letter ICAO code
\item 'lat'    : \code{numeric}    Latitude of airport
\item 'lon'    : \code{numeric}    Longitude of airport
}
}
\examples{
\dontrun{
ap = getAOI(state = "CO", county = "El Paso") \%>\% findAirports()
}
}
\author{
Mike Johnson
}
