% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGAGESII.R
\name{findGAGESII}
\alias{findGAGESII}
\title{Find GAGESII points and basins}
\usage{
findGAGESII(AOI = NULL, ids = FALSE, basins = FALSE)
}
\arguments{
\item{AOI}{and AOI of Interest}

\item{ids}{If TRUE, returns a list of COMIDS for NHD reaches}

\item{basins}{If TRUE, returns a list of GAGESII basin in addition}
}
\description{
\code{findNHD} returns a list of \code{Spatial*} Objects cropped to an Area of Interest.\cr\cr
To better understand defining an AOI using '\emph{state}', '\emph{county}' and '\emph{clip}' see \code{getAOI} and \code{getClipUnit}.\cr\cr
Returned \code{list} can be interactivly explored via \code{\link{explore}} and COMID values (\code{ids = TRUE}) allow for National Water Model access via \code{getNWM}.\cr\cr
All outputs are projected to \code{CRS'+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs'} and stream networks are (\emph{down})loaded from the \href{https://cida.usgs.gov}{USGS}.
}
\seealso{
Other HydroData 'find' functions: \code{\link{findAirports}},
  \code{\link{findGHCN}}, \code{\link{findHUC}},
  \code{\link{findNHD}}, \code{\link{findNID}},
  \code{\link{findNWIS}}, \code{\link{findRoads}},
  \code{\link{findSnotel}}, \code{\link{findWaterbodies}}
}
\author{
Mike Johnson
}
