% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGCHN.R
\name{findGCHN}
\alias{findGCHN}
\title{Locate all NOAA GCHN Stations within Area of Interest}
\usage{
findGCHN(state = NULL, county = NULL, clip_unit = NULL,
  get.boundary = FALSE, get.basemap = NULL, ids = FALSE, save = FALSE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{get.boundary}{logical. If TRUE, the AOI shapefile will be returned with station data in a list}

\item{get.basemap}{logical. If TRUE, the google basemap will be returned with gage station in a list}

\item{ids}{logical. If TRUE, return a vector of station ids in AOI. Can be passed to \code{getGCHN}}

\item{save}{logical. If TRUE, all data is written to a HydroData folder in the working directory}
}
\description{
Function to find GCHN stations within Area of Interest
}
\examples{
\dontrun{
#Find all GCHN stations in Harris County, Texas

harris.clim = findGCHN(state = 'TX', county = 'Harris',
              keep.boundary = TRUE, keep.basemap = TRUE, ids = TRUE, save = TRUE)
plot(harris.clim$basemap)
plot(harris.clim$boundary, add = TRUE, lwd = 5)
plot(harris.clim$noaa, add = TRUE, pch = 16, col = 'blue')

#Get Station IDs

harris.clim$ids
}

}
\author{
Mike Johnson
}
