% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGCHN.R
\name{findGHCN}
\alias{findGHCN}
\title{Find GHCN Stations within Area of Interest}
\usage{
findGHCN(state = NULL, county = NULL, clip_unit = NULL,
  boundary = FALSE, basemap = NULL, parameters = NULL, ids = FALSE,
  save = FALSE)
}
\arguments{
\item{state}{character. Provide full name(s) or two character abbriviation(s). Not case senstive}

\item{county}{character. Provide county name(s). Requires \code{state} input.}

\item{clip_unit}{SpatialObject* or list. For details see \code{?getClipUnit}}

\item{boundary}{logical. If TRUE, the AOI \code{SpatialPolygon(s)} will be joined to returned list}

\item{basemap}{logical. If TRUE, a basemap will be joined to returned list from \code{\link[dismo]{gmap}}.

If a user wants greater control over basemap apperance choose from:
\itemize{
\item't':  a terrain imagery basemap
\item's':  a sattilite imagery basemap
\item'h':  a hybrid imagery basemap
\item'r':  a roadmap imagery basemap
}}

\item{parameters}{character. Specifiy the variable of interest.The fice principle elemets are:\cr
PRCP = Precipitation (tenths of mm)\cr
SNOW = Snowfall (mm)\cr
SNWD = Snow depth (mm)\cr
TMAX = Maximum temperature (tenths of degrees C)\cr
TMIN = Minimum temperature (tenths of degrees C)\cr
\cr}

\item{ids}{logical. If TRUE, returns a list of COMIDS for all stream reaches in AOI}

\item{save}{logical. If TRUE, all data is saved to a HydroData folder created in users working directory. Find working directory with \code{\link}}
}
\value{
All HydroData outputs are projected to \emph{'+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs'}
}
\description{
\code{findGHCN} finds Global Historical Climatology Network (GHCN) stations within Area of Interest. Metadata allows for easy data download via
 the \code{getGHCN} function. To better understand how to define an AOI see \code{?getAOI}.
}
\details{
\code{findGHCN} returns a named list of minimum length 1:
\enumerate{
\item 'ghcn':      A \code{SpatialPointsDataFrame} of GHCN stations and metadata
\item 'basemap':   A \code{RasterLayer} basemap if 'basemap' is \code{TRUE}
\item 'fiat':      A \code{SpatialPolygon} of fiat boundaries if 'boundary' is \code{TRUE}
\item 'clip':      A \code{SpatialPolygon} of clip unit boundary if 'boundary' is \code{TRUE}
\item 'ids':       A vector of flowline COMIDs if 'ids' is \code{TRUE}
}
}
\examples{
\dontrun{
#Find all GHCN stations in Harris County, Texas

harris = findGHCN(state = 'TX',
                  county = 'Harris',
                  boundary = TRUE,
                  basemap = TRUE)

plot(harris$basemap)
plot(harris$fiat, add = TRUE, lwd = 5)
plot(harris$ghcn, add = TRUE, pch = 16, col = 'blue')

}

}
\seealso{
\itemize{
         \item \code{\link{getClipUnit}}
         \item \link{getAOI}
         \item \link[HydroData]{getGHCN}
         }

Other HydroData 'find' functions: \code{\link{findNHD}},
  \code{\link{findNID}}, \code{\link{findUSGS}}
}
\author{
Mike Johnson
}
\concept{HydroData 'find' functions}
