% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestAirports.R
\name{findNearestAirports}
\alias{findNearestAirports}
\title{Find Neareast Airports}
\usage{
findNearestAirports(point = NULL, n = 5, ids = FALSE, AOI = FALSE)
}
\arguments{
\item{n}{the number of Airports to return}

\item{location}{a lat, long pair}
}
\value{
A list contatining a data.frame of Airports and a minimum bounding box data.frame
}
\description{
Find Neareast Airports
}
\author{
Mike Johnson
}
