% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestCOMID.R
\name{findNearestCOMID}
\alias{findNearestCOMID}
\title{Find Neareast COMIDs}
\usage{
findNearestCOMID(point = NULL, n = 5, ids = FALSE, AOI = FALSE)
}
\arguments{
\item{location}{a lat, long pair or Colloqiual Name}
}
\value{
A list contatining a data.frame of Airports and a minimum bounding box data.frame
}
\description{
Find Neareast COMIDs
}
\author{
Mike Johnson
}
