% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestCOMID.R
\name{findNearestCOMID}
\alias{findNearestCOMID}
\title{Find Neareast COMIDs}
\usage{
findNearestCOMID(location = NULL, n = 5)
}
\arguments{
\item{location}{a lat, long pair}

\item{n}{the number of COMIDs to return}
}
\value{
A list contatining a data.frame of COMIDs and a minimum bounding box data.frame
}
\description{
Find Neareast COMIDs
}
\author{
Mike Johnson
}
