% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestGHCN.R
\name{findNearestGHCN}
\alias{findNearestGHCN}
\title{Find Neareast GHCN stations}
\usage{
findNearestGHCN(point = NULL, n = 5, parameters = NULL,
  ids = FALSE, bb = FALSE)
}
\arguments{
\item{point}{a point described by lat/long, can be piped from \link[AOI]{geocode}}

\item{n}{the number of stations to find (default = 5)}

\item{parameters}{Can be used to filter the returned stations by measurement types}

\item{ids}{If TRUE, a vector of airport stations IDs is added to retuned list (default = \code{FALSE})}

\item{bb}{If TRUE, the geometry of the minimum bounding area of the features is added to returned list  (default = \code{FALSE})}
}
\value{
a list() of minimum length 2: AOI and ghcn
}
\description{
\code{findNearestGHCN} returns a \code{SpatialPointsDataFrame} of the 'n' number of GHCN station closest to a declared point.
}
\examples{
\dontrun{
pt = geocode("UCSB") \%>\% findNearestGHCN(n = 5)
}
}
\seealso{
findGHCN
}
\author{
Mike Johnson
}
