% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestHUC12.R
\name{findNearestHUC12}
\alias{findNearestHUC12}
\title{Find Neareast COMIDs}
\usage{
findNearestHUC12(point = NULL, n = 1, ids = FALSE, AOI = FALSE)
}
\arguments{
\item{location}{a lat, long pair or Colloqiual Name}
}
\value{
A list contatining a data.frame of Airports and a minimum bounding box data.frame
}
\description{
Find Neareast COMIDs
}
\author{
Mike Johnson
}
