% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNED.R
\name{findNED}
\alias{findNED}
\title{Get NED data for an Area of Interest}
\usage{
findNED(AOI = NULL, res = 1)
}
\arguments{
\item{AOI}{a character string. Can be full name or state abbriviation}

\item{res}{resolution of NED data. 1 equals 1 arc second, 13 equals 1/3 arc second.}
}
\description{
Function to download elevation data for an area of interest
}
\examples{
\dontrun{
#Get 1 arc second elevation for El Paso County Colorado Springs

el.paso.elev = get_ned(state = "CO", county = "El Paso", res = 1, keep.boundary = TRUE)
plot(el.paso.elev, col = col_elev)
}
}
\author{
Mike Johnson
}
