% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNHD.R
\name{findNHD}
\alias{findNHD}
\title{Find National Hydrography Data Stream Networks}
\usage{
findNHD(state = NULL, county = NULL, clip_unit = NULL, boundary = FALSE,
  basemap = FALSE, ids = FALSE, save = FALSE)
}
\arguments{
\item{state}{Full name(s) or two character abbriviation(s). Not case senstive}

\item{county}{County name(s). Requires \code{state} input.}

\item{clip_unit}{SpatialObject* or list. For details see \code{getClipUnit}}

\item{boundary}{If TRUE, the AOI \code{SpatialPolygon(s)} will be joined to returned list}

\item{basemap}{If TRUE, a basemap will be joined to returned list

 If a user wants greater control over basemap apperance replace TRUE with either:
\itemize{
\item't':  google terrain basemap
\item's':  google sattilite imagery basemap
\item'h':  google hybrid basemap
\item'r':  google roads basemap
}}

\item{ids}{If TRUE, returns a list of COMIDS for NHD reaches}

\item{save}{If TRUE, data is written to a HydroData folder in users working directory.}
}
\value{
\code{findNHD} returns a list of minimum length 1:

\enumerate{
\item 'flowlines': A \code{SpatialLinesDataFrame}\cr


Pending parameterization, \code{findNHD} can also return:

\item 'basemap':   A \code{RasterLayer*} basemap if \code{basemap = TRUE}
\item 'boundry':   A \code{SpatialPolygon*} of AOI if \code{boundary = TRUE}
\item 'fiat':      A \code{SpatialPolygon*} of intersected county boundaries if \code{boundary = TRUE}
\item 'ids':       A vector of COMIDs if \code{ids = TRUE}
}
}
\description{
\code{findNHD} returns a list of \code{Spatial*} Objects cropped to an Area of Interest.\cr\cr
To better understand defining an AOI using '\emph{state}', '\emph{county}' and '\emph{clip_unit}' see \code{getAOI} and \code{getClipUnit}.\cr\cr
Returned \code{list} can be interactivly explored via \code{\link{explore}} and COMID values (\code{ids = TRUE}) allow for National Water Model access via \code{getNWM}.\cr\cr
All outputs are projected to \code{CRS'+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs'} and stream networks are (\emph{down})loaded from the \href{https://cida.usgs.gov}{USGS}.
}
\examples{
\dontrun{
# Find NHD data for El Paso County, Colorado

el.paso = findNHD(state = 'CO',
                  county = 'El Paso',
                  boundary = TRUE,
                  basemap = 'r',
                  ids = TRUE)

# Static Mapping

 plot(el.paso$basemap)
 plot(el.paso$boundary, lwd = 5, add = T)
 plot(el.paso$flowlines, col = 'blue', lwd = el.paso$flowlines$streamorde, add = T)

# Generate Interactive Map

  explore(el.paso, save = TRUE)

# Get flow data for reaches

flows = getNWM(ids = el.paso$ids, startDate = '2017-01-01', endDate = '2017-12-31')

# Interactivly explore timeseries data

inspect(flows,   save = TRUE)
}

}
\seealso{
\code{\link{getAOI}}

\code{\link{getNWM}}

\code{\link{explore}}

Other HydroData 'find' functions: \code{\link{findAirports}},
  \code{\link{findClosestAirports}},
  \code{\link{findGHCN}}, \code{\link{findNID}},
  \code{\link{findReservoir}}, \code{\link{findRoads}},
  \code{\link{findSnotel}}, \code{\link{findUSGS}},
  \code{\link{findWS}}, \code{\link{findWaterbodies}}
}
\author{
Mike Johnson
}
