% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findFlowlines.R
\name{findNHD}
\alias{findNHD}
\title{Find all NHD flowlines within an Area of Interest}
\usage{
findNHD(state = NULL, county = NULL, clip_unit = NULL,
  keep.boundary = FALSE, keep.basemap = FALSE, comids = FALSE,
  save = FALSE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with gage data in a list}

\item{keep.basemap}{logical. If TRUE, the google basemap will be returned with gage data in a list}

\item{comids}{logical. If TRUE, returns a list of COMIDS for all stream reaches on AOI}

\item{save}{logical. If TRUE, all data is written to a HydroData folder in the working directory}
}
\description{
Download NHD flowlines within an Area of Interest
}
\examples{
\dontrun{
#Find all NHD flowlines in El Paso County, Colorado

el.paso = findNHD(state = 'CO', county = 'El Paso',
          keep.boundary = TRUE, keep.basemap = TRUE, comids = TRUE, save = TRUE)

plot(el.paso$basemap)
plot(el.paso$boundary, add = TRUE, lwd = 5)
plot(el.paso$flowlines, add = TRUE, col = 'blue', lwd = el.paso$flowlines$streamorde)
}
}
\author{
Mike Johnson
}
