% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNID.R
\name{findNID}
\alias{findNID}
\title{Find US Army Core Dams within Area of Interest}
\usage{
findNID(state = NULL, county = NULL, clip_unit = NULL, boundary = FALSE,
  basemap = FALSE, save = FALSE)
}
\arguments{
\item{state}{character. Provide full name(s) or two character abbriviation(s). Not case senstive}

\item{county}{character. Provide county name(s). Requires \code{state} input.}

\item{clip_unit}{SpatialObject* or list. For details see \code{?getClipUnit}}

\item{boundary}{logical. If TRUE, the AOI \code{SpatialPolygon(s)} will be joined to returned list}

\item{basemap}{logical. If TRUE, a basemap will be joined to returned list from \code{\link[dismo]{gmap}}.

 #' If a user wants greater control over basemap apperance replace TRUE with either:
\itemize{
\item't':  a terrain imagery basemap
\item's':  a sattilite imagery basemap
\item'h':  a hybrid imagery basemap
\item'r':  a roadmap imagery basemap
}}

\item{save}{logical. If TRUE, all data is saved to a HydroData folder created in users working directory. Find working directory with \code{getwd()}}
}
\value{
All HydroData outputs are projected to \emph{'+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs'}
}
\description{
\code{findNID} finds all US Army Corps Dams for an Area of Interest from the National Inventory of Dams dataset.
This dataset is access through the dams package.

 \code{findNID} returns a named list of minimum length 1:
\enumerate{
\item 'dams':      A \code{SpatialPointsDataFrame} of NID dams and metadata
\item 'basemap':   A \code{RasterLayer} basemap if 'basemap' is \code{TRUE}
\item 'fiat':      A \code{SpatialPolygon} of fiat boundaries if 'boundary' is \code{TRUE}
\item 'clip':      A \code{SpatialPolygon} of clip unit boundary if 'boundary' is \code{TRUE}
}
}
\examples{
\dontrun{
# Find all dams in Texas

tx.dams = findNID(state = "TX", boundary = TRUE, basemap = 'r', save= TRUE)
plot(tx.dams$basmap)
plot(tx.dams$boundary, add = TRUE)
plot(tx.dams$dams, add = TRUE)
}
}
\seealso{
\itemize{
         \item \code{\link{getClipUnit}}
         \item \link{getAOI}
         \item \link[HydroData]{explore}
         }

Other HydroData 'find' functions: \code{\link{findAirports}},
  \code{\link{findClosestAirports}},
  \code{\link{findGHCN}}, \code{\link{findNHD}},
  \code{\link{findReservoir}}, \code{\link{findRoads}},
  \code{\link{findSnotel}}, \code{\link{findUSGS}},
  \code{\link{findWS}}, \code{\link{findWaterbodies}}
}
\author{
Mike Johnson
}
\concept{HydroData 'find' functions}
