% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNLCD.R
\name{findNLCD}
\alias{findNLCD}
\title{Find National Land Cover Products (NLCD)}
\usage{
findNLCD(AOI = NULL, year = 2011, type = "landcover")
}
\arguments{
\item{year}{year of land cover desired. Options include 2001, 2006, and 2011. Multiple years can be given as a vector.}

\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with station data in a list}
}
\description{
Function to download land cover data for an area of interest from 2001, 2006, 2011
}
\examples{
\dontrun{
el.paso.lc = get_nlcd_multi(state = 'CO', county = 'El Paso', year = c(2001, 2006, 2011))

plot(el.paso.lc$lc$lc_2001, col = col_lc$color, main = "2001")
plot(el.paso.lc$boundary, add = TRUE, lwd = 4)
legend('topright',col_lc$description,fill=col_lc$color, cex = .5)

plot(el.paso.lc$lc$lc_2006, col = col_lc$color, main = "2006")
plot(el.paso.lc$boundary, add = TRUE, lwd = 5)
legend('topright',col_lc$description,fill=col_lc$color, cex = .5)

plot(el.paso.lc$lc$lc_2011, col = col_lc$color, main = "2011")
plot(el.paso.lc$boundary, add = TRUE, lwd = 5)
legend('topright',col_lc$description,fill=col_lc$color, cex = .5)
}

}
\author{
Mike Johnson
}
