% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNWIS.R
\name{findNWIS}
\alias{findNWIS}
\title{Find USGS NWIS Stream Gages}
\usage{
findNWIS(AOI = NULL, ids = FALSE, comids = FALSE)
}
\arguments{
\item{ids}{If TRUE, returns a list of station IDs in AOI}

\item{comids}{If TRUE, returns a list of NHD COMIDs that correspond to the station IDs}

\item{state}{Full name(s) or two character abbriviation(s). Not case senstive}

\item{county}{County name(s). Requires \code{state} input.}

\item{clip}{SpatialObject* or list. For details see \code{getClipUnit}}

\item{boundary}{If TRUE, the AOI \code{SpatialPolygon(s)} will be joined to returned list

 If a user wants greater control over basemap apperance replace TRUE with either:
\itemize{
\item't':  google terrain basemap
\item's':  google sattilite imagery basemap
\item'h':  google hybrid basemap
\item'r':  google roads basemap
}}

\item{save}{If TRUE, data is written to a HydroData folder in users working directory.}
}
\value{
\code{findUSGS} returns a named list of minimum length 1:

\enumerate{
\item 'nwis':     A \code{SpatialPointsDataFrame} of stations and metadata
Pending parameterization, \code{findRoads} can also return:

\item 'basemap':   A \code{RasterLayer*} basemap if \code{basemap = TRUE}
\item 'boundry':   A \code{SpatialPolygon*} of AOI if \code{boundary = TRUE}
\item 'fiat':      A \code{SpatialPolygon*} of intersected county boundaries if \code{boundary = TRUE}
\item 'ids':       A vector of road names if \code{ids = TRUE}
}
}
\description{
\code{findUSGS} returns a list of \code{Spatial*} Objects cropped to an Area of Interest.\cr\cr
To better understand defining an AOI using '\emph{state}', '\emph{county}' and '\emph{clip}' see \code{getAOI} and \code{getClipUnit}.\cr\cr
Returned \code{list} can be interactivly explored via \code{\link{explore}} \cr
Returned ID values (\code{ids = TRUE}) allow for observed data access via \code{getUSGS} or \link[dataRetrieval]{readNWISdv} .\cr
Returned COMID values (\code{comids = TRUE}) allow for National Water Model access via \code{getNWM}.\cr\cr

All outputs are projected to \code{CRS '+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs'} and station data was extracted from the \href{https://maps.waterdata.usgs.gov/mapper/index.html}{USGS NWIS Mapper}.
}
\examples{
\dontrun{
# Find stations within 10 square miles of the National Water Center:

 nwc.usgs <- findUSGS(clip = list("National Water Center", 10,  10),
                             boundary = TRUE, basemap = TRUE, ids = TRUE)

# Static Mapping

 plot(nwc.usgs$basemap)
 plot(nwc.usgs$boundary, add = TRUE, lwd = 5)
 plot(nwc.usgs$usgs, add = TRUE, lwd = 2, pch = 16,  col = "darkgreen")

# Generate Interactive Map

explore(nwc.usgs)

# Get discharge data for all stations usign dataRetrival Package

 data = getUSGS(IDs = nwc.usgs$ids)

# Generate Interactive Timseries

inspect(data, param = 'Flow')
}

}
\seealso{
\code{\link{getAOI}}

\code{\link{explore}}

\code{\link{getUSGS}}

\code{\link[dataRetrieval]{readNWISdv}}

Other HydroData 'find' functions: \code{\link{findAirports}},
  \code{\link{findGAGESII}}, \code{\link{findGHCN}},
  \code{\link{findHUC}}, \code{\link{findNHD}},
  \code{\link{findNID}}, \code{\link{findRoads}},
  \code{\link{findSnotel}}, \code{\link{findWaterbodies}}
}
\author{
Mike Johnson
}
