% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findRoads.R
\name{findRoads}
\alias{findRoads}
\title{Find TIGER roads within Area of Interest}
\usage{
findRoads(state = NULL, county = NULL, clip_unit = NULL,
  keep.boundary = FALSE, keep.basemap = FALSE, save = FALSE)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}

\item{keep.boundary}{logical. If TRUE, the AOI shapefile will be returned with gage data in a list}

\item{keep.basemap}{logical. If TRUE, the google basemap will be returned with gage data in a list}

\item{save}{logical. If TRUE, all data is written to a HydroData folder in the working directory}
}
\description{
Function to locate all TIGER roads within an Area of Interest from the US Census Bureau
}
\examples{
\dontrun{
ep.roads = find_roads(state = 'CO', county = 'el paso', keep.boundary = TRUE, keep.basemap = TRUE)
plot(ep.roads$basemap)
plot(ep.roads$boundary, add = T, lwd = 5 )
plot(ep.roads$roads,    add = T )
}

}
\author{
Mike Johnson
}
