% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSnotel.R
\name{findSnotel}
\alias{findSnotel}
\title{Find USDA NRCS Snotel Stations}
\usage{
findSnotel(AOI = NULL, ids = FALSE)
}
\arguments{
\item{ids}{If TRUE, returns a list of station IDs in AOI}

\item{state}{Full name(s) or two character abbriviation(s). Not case senstive}

\item{county}{County name(s). Requires \code{state} input.}

\item{clip}{SpatialObject* or list. For details see \code{getClipUnit}}

\item{boundary}{If TRUE, the AOI \code{SpatialPolygon(s)} will be joined to returned list

 If a user wants greater control over basemap apperance replace TRUE with either:
\itemize{
\item't':  google terrain basemap
\item's':  google sattilite imagery basemap
\item'h':  google hybrid basemap
\item'r':  google roads basemap
}}

\item{save}{If TRUE, data is written to a HydroData folder in users working directory.}
}
\value{
\code{findSnotel} returns a list of minimum length 1:

\enumerate{
\item 'snotel': A \code{SpatialPointsDataFrame*}\cr

Pending parameterization, \code{findSnotel} can also return:

\item 'basemap':   A \code{RasterLayer*} basemap if \code{basemap = TRUE}
\item 'boundry':   A \code{SpatialPolygon*} of AOI if \code{boundary = TRUE}
\item 'fiat':      A \code{SpatialPolygon*} of intersected county boundaries if \code{boundary = TRUE}
\item 'ids':       A vector of station IDs if \code{ids = TRUE}
}
}
\description{
\code{findSnotel} returns a list of \code{Spatial*} Objects cropped to an Area of Interest.\cr\cr
To better understand defining an AOI using '\emph{state}', '\emph{county}' and '\emph{clip}' see \code{getAOI} and \code{getClipUnit}.\cr\cr
Returned \code{list} can be interactivly explored via \code{\link{explore}} and ID values (\code{ids = TRUE}) allow for SNOTEL data access via \code{getSNOTEL}.\cr\cr
All outputs are projected to \code{CRS '+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs'} and station data is taken from \href{https://wcc.sc.egov.usda.gov/nwcc/yearcount?network=sntl&counttype=statelist&state=}{NRCS reports}.
}
\examples{
\dontrun{
# Find Snotel stations in Nevada

nv.snow = findSnotel(state = 'NV', boundary = TRUE, basemap = TRUE)

# Static Mapping

plot(nv.snow$basemap)
plot(nv.snow$boundary, add = TRUE, lwd = 5)
plot(nv.snow$snotel, add = TRUE, pch = 8, col = "lightblue")

# Generate Interactive Map

explore(nv.snow)
}


}
\seealso{
\code{\link{getAOI}}

\code{\link{getSnotel}}

\code{\link{explore}}

Other HydroData 'find' functions: \code{\link{findHUC}},
  \code{\link{findNWIS}}, \code{\link{findRoads}},
  \code{\link{findWaterbodies}}
}
\author{
Mike Johnson
}
