% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSnotel.R
\name{findSnotel}
\alias{findSnotel}
\title{Find USDA NRCS Snotel Stations}
\usage{
findSnotel(AOI = NULL, ids = FALSE)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{ids}{If TRUE,  a vector of NHD COMIDs is added to retuned list (default = \code{FALSE})}
}
\description{
\code{findSnotel} returns a \code{SpatialPointsDataFrame*} Objects cropped to an Area of Interest.
Station data is extacted from \href{https://wcc.sc.egov.usda.gov/nwcc/yearcount?network=sntl&counttype=statelist&state=}{NRCS reports} and contains 11 attributes:
\itemize{
\item 'NETWORK'   : \code{integer}  Description of Network (all = 'SNTL')
\item 'STATE'   : \code{character}  The state the station is located in
\item 'NAME': \code{POSITct}  Unique station name
\item 'ID'   : \code{character}    Unique identifier assigned by NRCS
\item 'START.DATE'   : \code{character}    Date the station made first measurment ("Year-Month")
\item 'LAT'    : \code{numeric}    Latitude of station, decimil degrees
\item 'LON'    : \code{numeric}    Longitude of station, decimil degrees
\item 'ELEV'   : \code{character}     Elevation of station
\item 'COUNTY'   : \code{integer}  The county the station is located in
\item 'HUC12.NAME': \code{character}  The HUC12 name the station is located in
\item 'HUC12.ID': \code{character}  The HUC12 ID the station is located in
}\cr
}
\examples{
\dontrun{
CA.sno = getAOI(state = 'CA') \%>\% findSnotel()
}
}
\author{
Mike Johnson
}
