% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWaterbodies.R
\name{findWaterbodies}
\alias{findWaterbodies}
\title{Find USA Water Bodies (Polygon)}
\usage{
findWaterbodies(state = NULL, county = NULL, clip_unit = NULL,
  boundary = FALSE, basemap = FALSE, ids = FALSE, save = FALSE)
}
\arguments{
\item{state}{Full name(s) or two character abbriviation(s). Not case senstive}

\item{county}{County name(s). Requires \code{state} input.}

\item{clip_unit}{SpatialObject* or list. For details see \code{getClipUnit}}

\item{boundary}{If TRUE, the AOI \code{SpatialPolygon(s)} will be joined to returned list}

\item{basemap}{If TRUE, a basemap will be joined to returned list

 If a user wants greater control over basemap apperance replace TRUE with either:
\itemize{
\item't':  google terrain basemap
\item's':  google sattilite imagery basemap
\item'h':  google hybrid basemap
\item'r':  google roads basemap
}}

\item{ids}{If TRUE, returns a list of water body names in AOI}

\item{save}{If TRUE, data is written to a HydroData folder in users working directory.}
}
\value{
\code{findWaterbodies} returns a named list of minimum length 1:

\enumerate{
\item 'waterbodies': A \code{SpatialLinesDataFrame*}\cr

Pending parameterization, \code{findRoads} can also return:

\item 'basemap':   A \code{RasterLayer*} basemap if \code{basemap = TRUE}
\item 'boundry':   A \code{SpatialPolygon*} of AOI if \code{boundary = TRUE}
\item 'fiat':      A \code{SpatialPolygon*} of intersected county boundaries if \code{boundary = TRUE}
\item 'ids':       A vector of waterbody names if \code{ids = TRUE}
}
}
\description{
\code{findWaterboides} returns a list of \code{Spatial*} Objects cropped to an Area of Interest.\cr\cr
To better understand defining an AOI using '\emph{state}', '\emph{county}' and '\emph{clip_unit}' see \code{getAOI} and \code{getClipUnit}.\cr\cr
Returned \code{list} can be interactivly explored via \code{\link{explore}}.\cr\cr
All outputs are projected to \code{CRS '+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs'} and spatial water bodies data is (\emph{down})loaded from the \href{http://open-esrifederal.opendata.arcgis.com/datasets/9dff3cf646704abd9e74265f02abeb09_0}{ESRI Federal Open Data Portal}.
}
\examples{
\dontrun{
# Find  water bodies in San Diego County, CA

wb = findWaterbodies(state = "CA", county = "San Diego", basemap = T, boundary = T)

# Static Mapping

plot(wb$basemap)
plot(wb$boundary, add = T)
plot(wb$waterbodies, add = T, col = 'blue')

# Generate Interactive Map

explore(wb)
}
}
\seealso{
\code{\link{getAOI}}

\code{\link{findReservoir}}

\code{\link{explore}}

Other HydroData 'find' functions: \code{\link{findAirports}},
  \code{\link{findClosestAirports}},
  \code{\link{findGHCN}}, \code{\link{findHUC}},
  \code{\link{findNHD}}, \code{\link{findNID}},
  \code{\link{findReservoir}}, \code{\link{findRoads}},
  \code{\link{findSnotel}}, \code{\link{findUSGS}}
}
\author{
Mike Johnson
}
