% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWS.R
\name{findWS}
\alias{findWS}
\title{Find Watershed Boundaries (WBD)}
\usage{
findWS(state = NULL, county = NULL, clip_unit = NULL, boundary = FALSE,
  basemap = FALSE, ids = FALSE, level = 10, subbasins = FALSE,
  HUC8 = NULL, save = FALSE)
}
\arguments{
\item{state}{Full name(s) or two character abbriviation(s). Not case senstive}

\item{county}{County name(s). Requires \code{state} input.}

\item{clip_unit}{SpatialObject* or list. For details see \code{getClipUnit}}

\item{boundary}{If TRUE, the AOI \code{SpatialPolygon(s)} will be joined to returned list}

\item{basemap}{If TRUE, a basemap will be joined to returned list

 If a user wants greater control over basemap apperance replace TRUE with either:
\itemize{
\item't':  google terrain basemap
\item's':  google sattilite imagery basemap
\item'h':  google hybrid basemap
\item'r':  google roads basemap
}}

\item{ids}{If TRUE, returns a list of HUC8 in AOI}

\item{level}{defines the HUC level of interest}

\item{subbasins}{If TRUE, all subbasins of the supplied \code{level} will be joined to retuned list}

\item{HUC8}{option to supply known HUC8 code(s)}

\item{save}{If TRUE, data is written to a HydroData folder in users working directory.}
}
\value{
\code{findWS} returns a named list of minimum length 1:

\enumerate{
\item 'huc_': A \code{SpatialPolygonDataFrame*} for each requested HUC level defined by \code{level} and \code{subbasins} \cr

Pending parameterization, \code{findWS} can also return:

\item 'basemap':   A \code{RasterLayer*} basemap if \code{basemap = TRUE}
\item 'boundry':   A \code{SpatialPolygon*} of AOI if \code{boundary = TRUE}
\item 'fiat':      A \code{SpatialPolygon*} of intersected county boundaries if \code{boundary = TRUE}
\item 'ids':       A vector of HUC8 units if \code{ids = TRUE}
}
}
\description{
\code{findWS} returns a list of \code{Spatial*} Objects cropped to an Area of Interest.\cr\cr
To better understand defining an AOI using '\emph{state}', '\emph{county}' and '\emph{clip_unit}' see \code{getAOI} and \code{getClipUnit}.\cr\cr
Returned \code{list} can be interactivly explored via \code{\link{explore}}.\cr\cr
All outputs are projected to \code{CRS '+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs'} and spatial watershed data are (\emph{down})loaded from the  \href{https://prd-tnm.s3.amazonaws.com}{USGS Staged Data Server}.
}
\examples{
\dontrun{
# Find all HUC8, 10 and 12 units surrounding the National Water Center

 ws = findWS(clip_unit = list("National Water Center", 10 ,10),
            boundary = T,
            basemap = T,
            subbasins = T,
            level = 8)

# Static Mapping

plot(ws$basemap)
plot(ws$boundary, add = T)
plot(ws$huc_8, lwd = 3, add = T)
plot(ws$huc_10, border = 'red', lwd = 2, add = T)
plot(ws$huc_12, border = 'blue', add = T)

# Generate Interactive Map

explore(ws)
}

}
\seealso{
\code{\link{getAOI}}

\code{\link{explore}}

Other HydroData 'find' functions: \code{\link{findAirports}},
  \code{\link{findClosestAirports}},
  \code{\link{findGHCN}}, \code{\link{findNHD}},
  \code{\link{findNID}}, \code{\link{findReservoir}},
  \code{\link{findRoads}}, \code{\link{findSnotel}},
  \code{\link{findUSGS}}, \code{\link{findWaterbodies}}
}
\author{
Mike Johnson
}
\concept{HydroData 'find' functions}
