% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_sp.R
\name{get_region_sp}
\alias{get_region_sp}
\title{A function for finding County and State shapefiles}
\usage{
get_region_sp(state = NULL, county = NULL,
  proj = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{proj}{a projction string. Default is a EPSG:4326}
}
\value{
A character string
}
\description{
This function utilizes the Maptools package to generate a shapefile for a county or state given a user query.
}
\examples{
get_region_sp(state = "CA", county = "San Luis Obispo")

}
\author{
Mike Johnson
}
