% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFiatBoundary.R
\name{getFiatBoundary}
\alias{getFiatBoundary}
\title{Get County and State SpatialPolygon(s)}
\usage{
getFiatBoundary(state = NULL, county = NULL, clip_unit = NULL)
}
\arguments{
\item{state}{character. Full name or two character abbriviation. Not case senstive}

\item{county}{character. Provide county name(s). Requires 'state' input.}

\item{clip_unit}{SpatialObject* or list. For details see \code{?getClipUnit}}
}
\value{
\code{getFiatBoundary} returns a \code{SpatialPolygon} Object
}
\description{
Get County and State SpatialPolygon(s)
}
\details{
\code{getFiatBoundary} gets a \code{SpatialPolygon} for a defiend state and/or county or those intersecting a clip_unit.
Fiat boundaries come from the 2017 US Census Bureau 2017 TIGER Dataset.

All HydroData outputs are projected to \emph{'+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs'}
}
\examples{
\dontrun{
# Get Single State
    getFiatBoundary(state = "CA")

# Get Multi-state
    getFiatBoundary(state = c("CA","Utah","Nevada"))

# Get County
    getFiatBoundary(state = "CA", county = "San Luis Obispo")

# Get Muli-county
   getFiatBoundary(state = "CA", county = c("San Luis Obispo", "Santa Barbara", "Ventura"))

# Get counties that intersect with defined clip_unit
   getFiatBoundary(clip_unit = list("UCSB", 10, 10, "lowerleft"))
}

}
\seealso{
\itemize{
         \item \code{\link{getClipUnit}}
         \item \code{\link{getAOI}}
         }

Other HydroData 'get' functions: \code{\link{getAOI}},
  \code{\link{getClipUnit}}, \code{\link{getDaymet_pixel}},
  \code{\link{getUSGS}}
}
\author{
Mike Johnson
}
