% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HydroData.R
\docType{package}
\name{HydroData}
\alias{HydroData}
\alias{HydroData-package}
\title{\code{HydroData} package}
\description{
Access climate, landscape and hydrolgic data programatically
}
\details{
See the README on
#href{https://cran.r-project.org/package=googlesheets/README.html}{CRAN}
#href{https://github.com/jennybc/googlesheets#readme}{GitHub}
}
