% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{inspect}
\alias{inspect}
\title{Inspect Timeseries data found with HydroData}
\usage{
inspect(data = NULL, timestep = "daily", param = NULL, save = FALSE)
}
\arguments{
\item{data}{data source to visualize}

\item{timestep}{time step to aggregate to, options include "daily", "monthly", "yearly"}

\item{param}{the parameter to visualize, must match a column name in the data input}

\item{save}{should the HTML dygraph be save to the users disk. Default FALSE}
}
\description{
\code{inspect} can be used to explore timeseries HydroData objects
Namely:
 getUSGS
 getWU
 getNWM
 getGHCN
 getDaymet_point
}
\seealso{
Other HydroData 'viz' functions: \code{\link{explore}}
}
\author{
Mike Johnson
}
\concept{HydroData 'viz' functions}
