% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameAOI.R
\name{nameAOI}
\alias{nameAOI}
\title{A function for returning a text string name of AOI}
\usage{
nameAOI(state = NULL, county = NULL, clip_unit = NULL)
}
\arguments{
\item{state}{a character string. Can be full name or state abbriviation}

\item{county}{a character string. Can be full name or state abbriviation}

\item{clip_unit}{can be provided as a shapefile or as a vector defineing centroid and bounding box diminsion}
}
\value{
A character string
}
\description{
A function for returning a text string name of AOI
}
\examples{
\dontrun{
nameAOI(state = "CA")
}
}
\author{
Mike Johnson
}
