% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_return_what.R
\name{return.what}
\alias{return.what}
\title{Choose what to return for HydroData Calls}
\usage{
return.what(sp, items, report, AOI, boundary, clip, ids)
}
\arguments{
\item{sp}{Spatial object}

\item{items}{an items list}

\item{report}{a report vector}

\item{AOI}{the defined AOI}

\item{boundary}{the boundary call}

\item{clip}{the clip defintion}

\item{ids}{the ids call}
}
\value{
a list of HydroData components
}
\description{
A function defining what should be returned in a HydroData object
}
\author{
Mike Johnson
}
