% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_what.R
\name{return.what}
\alias{return.what}
\title{Choose what to return for HydroData Calls}
\usage{
return.what(items, report, AOI, basemap, boundary, clip_unit, ids)
}
\arguments{
\item{items}{an items list}

\item{report}{a report vector}

\item{AOI}{the defined AOI}

\item{basemap}{the basemap call}

\item{boundary}{the boundary call}

\item{clip_unit}{the clip_unit defintion}

\item{ids}{the ids call}
}
\description{
A function defining what should be returned in a HydroData object
}
\seealso{
Other HydroData 'helper' functions: \code{\link{define.clip.unit}},
  \code{\link{name.clip.unit}}
}
\author{
Mike Johnson
}
\concept{HydroData 'helper' functions}
