% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_return_what.R
\name{return.what}
\alias{return.what}
\title{Choose what to return for HydroData Calls}
\usage{
return.what(AOI, type, report, vals)
}
\arguments{
\item{AOI}{Spatial object}

\item{report}{a report vector}

\item{vals}{the ids call}

\item{AOI}{the defined AOI}
}
\value{
a list of HydroData components
}
\description{
A function defining what should be returned in a HydroData object
}
\author{
Mike Johnson
}
