% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCap.R
\name{simpleCap}
\alias{simpleCap}
\title{A function for capitilizing all words in a string}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{a string to be parsed.}
}
\description{
A function for capitilizing all words in a string
}
\examples{
simpleCap("santa barbara")

}
\author{
Mike Johnson
}
