% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_writeHD.R
\name{writeHD}
\alias{writeHD}
\title{Write Data to disk}
\usage{
writeHD(class, dir, feature_name, ext = "shp")
}
\arguments{
\item{class}{objects class}

\item{dir}{directory}

\item{feature_name}{object id}

\item{ext}{shp or gpkg}
}
\description{
Write Data to disk
}
